% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functions.R
\name{activate_sink}
\alias{activate_sink}
\alias{deactivate_sink}
\title{Activate and deactivate sink}
\usage{
activate_sink(log_file_name)

deactivate_sink()
}
\arguments{
\item{log_file_name}{the name of the log file in which to sink the output (including the file path)}
}
\description{
A convenience function for programs that run at scale. activate_sink activates the sink to the log file, and deactivate_sink deactivates the sink.
}
