% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functs.R
\name{add_extra_covariates}
\alias{add_extra_covariates}
\title{Add extra covariates to a sceptre object}
\usage{
add_extra_covariates(sceptre_object, extra_covariates)
}
\arguments{
\item{sceptre_object}{TBD}

\item{extra_covariates}{TBD}
}
\value{
TBD
}
\description{
Add extra covariates to a sceptre object
}
