% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functs.R
\name{combine_perturbations}
\alias{combine_perturbations}
\title{Combine perturbations}
\usage{
combine_perturbations(perturbation_matrix, gRNA_groups_table)
}
\arguments{
\item{perturbation_matrix}{a perturbation matrix (i.e., a binary matrix of gRNA-to-cell assignments) stored as a sparse matrix (as implemented by the Matrix package) or a dense matrix (as implemented by base R); the row names should be the gRNA IDs}

\item{gRNA_groups_table}{a data frame with columns \code{gRNA_id} and \code{gRNA_group}, mapping each gRNA to its gRNA group}
}
\value{
a "combined" perturbation matrix, where gRNAs within the same gRNA group have been collapsed into a single row
}
\description{
Combine perturbations
}
\examples{
library(Matrix)
data("gRNA_matrix")
data("gRNA_groups_table")
perturbation_matrix <- threshold_gRNA_matrix(gRNA_matrix)
combined_perturbation_matrix <- combine_perturbations(perturbation_matrix, gRNA_groups_table)
}
