% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{compare_calibration_and_discovery_results}
\alias{compare_calibration_and_discovery_results}
\title{Title}
\usage{
compare_calibration_and_discovery_results(
  calibration_result,
  discovery_result,
  p_thresh,
  transform_scale = TRUE,
  include_legend = FALSE,
  include_y_axis_text = TRUE,
  point_size = 0.55,
  transparency = 0.8
)
}
\arguments{
\item{point_size}{(optional; default \code{0.55}) the size of the individual points in the plot.}

\item{transparency}{(optional; default \code{0.8}) the transparency of the individual points in the plot.}
}
\description{
Title
}
