% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{compute_cell_covariates}
\alias{compute_cell_covariates}
\title{Compute cell covariates}
\usage{
compute_cell_covariates(matrix_in, feature_names, compute_p_mito)
}
\arguments{
\item{matrix_in}{a matrix with features (e.g., genes, gRNAs, or proteins) in the rows and cells in the columns. The matrix can be a standard (dense) matrix or a sparse matrix of class \code{dgCMatrix}, \code{dgRMatrix}, or \code{dgTMatrix}. Row names are optional.}
}
\value{
a data frame containing the following cell covariates.
\itemize{
\item{\code{n_nonzero}}: the number of features expressed in the cell
\item{\code{n_umi}}: the total number of UMIs sequenced in the cell
\item{\code{p_mito}}: the fraction of UMIs sequenced in the cell that maps to mitochondrial genes
}
}
\description{
The \code{compute_cell_covariates()} function takes as input a feature-by-cell matrix and computes the cell covariates for that matrix.
}
\note{
The variable \code{p_mito} is computed only if feature names are provided and some feature names begin with "MT-"; these features are assumed to be mitochondrial genes.
}
\examples{
data(response_matrix_lowmoi)
cell_covariates <- compute_cell_covariates(response_matrix_lowmoi)
}
