% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_constructor_functs.R
\name{construct_all_pairs}
\alias{construct_all_pairs}
\title{Return all pairs}
\usage{
construct_all_pairs(
  sceptre_object,
  grna_groups_to_exclude = character(),
  response_grna_group_pairs_to_exclude = data.frame()
)
}
\arguments{
\item{grna_groups_to_exclude}{}
}
\description{
Return all pairs
}
