% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_constructor_functs.R
\name{construct_cis_pairs}
\alias{construct_cis_pairs}
\title{Construct cis pairs}
\usage{
construct_cis_pairs(
  sceptre_object,
  distance_threshold = 500000L,
  positive_control_pairs = data.frame(),
  ref_genome = "10X_GRCh38_2020"
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{distance_threshold}{TBD}

\item{positive_control_pairs}{TBD}

\item{ref_genome}{TBD}

\item{exclude_positive_control_grna_targets}{TBD}
}
\value{
TBD
}
\description{
\code{construct_cis_pairs()} is a helper function to construct the \emph{cis} pairs.
}
