% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lowmoi.R
\docType{data}
\name{covariate_data_frame_lowmoi}
\alias{covariate_data_frame_lowmoi}
\title{Covariate data frame}
\format{
An object of class \code{data.frame} with 20729 rows and 4 columns.
}
\usage{
data(covariate_data_frame_lowmoi)
}
\description{
A data frame containing the cell-specific technical factors:
\describe{
\item{\code{response_n_umis}}{(numeric) the total number of gene transcripts sequenced in the cell}
\item{\code{response_n_nonzero}}{(numeric) the total number of genes expressed in the cell}
\item{\code{bio_rep}}{(factor) the biological replicate in which the cell was sequenced (\code{rep_1}, \code{rep_2}, or \code{rep_3})}
\item{\code{p_mito}}{(numeric) the fraction of transcripts mapping to mitochondrial genes}
}
}
\section{Overview}{

The example low MOI data in the \code{sceptre} package are taken from the paper "Characterizing the molecular regulation of inhibitory immune checkpoints with multimodal single-cell screens" by Papalexi et al., 2021. The authors used CRISPRko to target 26 genes in 20,729 THP-1 cells and measured the effects of these perturbations via single-cell sequencing. The authors designed three or four gRNAs to target each gene, as well as a library of nine non-targeting gRNAs. The example data include 290 (or 2\\%) randomly selected genes from the original dataset.
}

\keyword{datasets}
