% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covariate_matrix}
\alias{covariate_matrix}
\title{Covariate matrix}
\format{
An object of class \code{data.frame} with 106660 rows and 4 columns.
}
\usage{
data(covariate_matrix)
}
\description{
A matrix of cell-specific technical factors:
\describe{
\item{\code{lg_gene_lib_size}}{(numeric) log-transformed gene library size}
\item{\code{lg_gRNA_lib_size}}{(numeric) log-transformed gRNA library size}
\item{\code{batch}}{(factor) sequencing batch (batch_1, ..., batch_5)}
\item{\code{p_mito}}{(numeric) fraction of sequenced gene transcripts that map to mitochondrial genes}
}
}
\section{Overview}{

This object is part of the example data in the \code{sceptre} package. The data are taken from the paper "Global Analysis of Enhancer Targets Reveals Convergent Enhancer-Driven Regulatory Modules" by Xie et al., 2019. The authors used CRISPRi-based perturb-seq to target >500 putative enhancers in a population of K562 cells. The example data consist of a subset of the sequenced gRNAs and genes measured on 106,660 cells.
}

\keyword{datasets}
