% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functions.R
\name{create_dictionary}
\alias{create_dictionary}
\title{Create dictionary}
\usage{
create_dictionary(ids, pod_size)
}
\arguments{
\item{ids}{the names of elements (gRNAs or genes)}

\item{pod_size}{size of the pods}

\item{precomp_base_filename}{base name of the file (specified with full file path) in which to store the precomputation.}
}
\description{
Create a dictionary that maps named elements (gRNAs, genes, or gRNA pairs) to files for a given "pod" size.
}
