% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_data_functions.R
\name{create_fbm_from_mtx}
\alias{create_fbm_from_mtx}
\title{Create FBM from .mtx file}
\usage{
create_fbm_from_mtx(path_to_mtx, dest_folder = NULL)
}
\arguments{
\item{path_to_mtx}{a file path to the .mtx file}

\item{dest_folder}{the folder in which to store the FBMs (default -- directory of .mtx file).}
}
\value{
the full file paths to the backing file (cell-gene matrix) and transposed backing file (gene-cell matrix).
}
\description{
Create FBM from .mtx file
}
\examples{
path_to_mtx <- "/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/raw/GSE120861_at_scale_screen.exprs.mtx"
dest_folder <- "/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/processed"
}
