% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_data_functions.R
\name{create_parent_directories}
\alias{create_parent_directories}
\title{Create parent directories}
\usage{
create_parent_directories(s)
}
\arguments{
\item{s}{a character string containing slashes}
}
\value{
ex) a/b/c returns a/b/c, a/b, a.
}
\description{
Create parent directories
}
