% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functs.R
\name{create_perturbation_matrix}
\alias{create_perturbation_matrix}
\title{Create perturbation matrix}
\usage{
create_perturbation_matrix(gRNA_matrix, gRNA_groups_table, threshold = 3)
}
\arguments{
\item{gRNA_matrix}{a gRNA-by-cell expression matrix stored as a sparse matrix (as implemented by the Matrix package) or a dense matrix (as implemented by base R); the row names should be the gRNA IDs}

\item{gRNA_groups_table}{a data frame with columns \code{gRNA_id} and \code{gRNA_group}.}

\item{threshold}{}
}
\value{

}
\description{
A threshold of 3 is a reasonable default for high-MOI single-cell CRISPR screens; see "Exponential family measurement error models for single-cell CRISPR screens" by Barry et al. 2022 for an in-depth analysis of thresholding and more sophisticated thresholding techniques.
}
\examples{
data("gRNA_matrix")
data("gRNA_groups")
gRNA_groups_table <- gRNA_groups
}
