% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_highmoi.R
\docType{data}
\name{discovery_pairs_highmoi}
\alias{discovery_pairs_highmoi}
\title{Discovery gene-gRNA group pairs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 358 rows and 2 columns.
}
\usage{
data(discovery_pairs_highmoi)
}
\description{
The discovery response-gRNA group pairs that we seek to test for association, with columns \code{response_id} (required), \code{grna_group} (required), and \code{type} (optional).
}
\section{Overview}{

This object is part of the example data. The data are taken from the paper "A genome-wide framework for mapping gene regulation via cellular genetic screens" by Gasperini et al., 2019. The authors used a CRISPRi-based assay to target 5,000+ putative enhancers in a population of K562 cells. The authors additionally targeted 200+ gene transcription start sites (TSSs) and designed a library of 50 non-targeting gRNAs to serve as negative controls. Genes, gRNAs, and cells are all down-sampled to reduce the size of the data.
}

\keyword{datasets}
