% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_matrix}
\alias{gene_matrix}
\title{Gene expression matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 106660 columns.
}
\usage{
data(gene_matrix)
}
\description{
A gene-by-cell expression matrix.
}
\section{Overview}{

This object is part of the example data in the \code{sceptre} package. The data are taken from the paper "Global Analysis of Enhancer Targets Reveals Convergent Enhancer-Driven Regulatory Modules" by Xie et al., 2019. The authors used a CRISPRi-based perturb-seq assay to target >500 putative enhancers in a population of K562 cells. The example data consist of a subset of the sequenced gRNAs and genes measured on 106,660 cells.
}

\keyword{datasets}
