% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_highmoi.R
\docType{data}
\name{gene_names_highmoi}
\alias{gene_names_highmoi}
\title{gene names high MOI}
\format{
An object of class \code{character} of length 526.
}
\usage{
data(gene_names_highmoi)
}
\description{
A vector giving the human-readable "name" of each gene.
}
\section{Overview}{

This object is part of the example data. The data are taken from the paper "A genome-wide framework for mapping gene regulation via cellular genetic screens" by Gasperini et al., 2019. The authors used a CRISPRi-based assay to target 5,000+ putative enhancers in a population of K562 cells. The authors additionally targeted 200+ gene transcription start sites (TSSs) and designed a library of 50 non-targeting gRNAs to serve as negative controls. Genes, gRNAs, and cells are all down-sampled to reduce the size of the data.
}

\keyword{datasets}
