% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_position_data_frame_grch37}
\alias{gene_position_data_frame_grch37}
\title{Gene position data frame (GRCh37)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 57773 rows and 4 columns.
}
\usage{
data(gene_position_data_frame_grch37)
}
\description{
\code{gene_position_data_frame_grch37} contains the coordinate and transcription start site position of each gene relative to reference genome GRCh37. It was constructed from the reference genome available on the 10x Genomics website.
}
\examples{
data(gene_position_data_frame_grch37)
head(gene_position_data_frame_grch37)
}
\keyword{datasets}
