% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{generate_all_pairs}
\alias{generate_all_pairs}
\title{Generate all pairs}
\usage{
generate_all_pairs(sceptre_object)
}
\value{
a data frame containing the columns \code{response_id} and \code{grna_group} in which each response is mapped to the entire set of targeting gRNA groups.
}
\description{
The \code{generate_all_pairs()} function generates all possible response-gRNA group pairs that can be constructed from a given \code{response_matrix} and \code{grna_group_data_frame}.
}
