% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomp_functs_v2.R
\name{generate_synthetic_grna_data}
\alias{generate_synthetic_grna_data}
\title{Generate synthetic data}
\usage{
generate_synthetic_grna_data(fitted_probs, B)
}
\arguments{
\item{fitted_probs}{a vector of fitted grna presence/absence indicators}

\item{B}{the number of resampled vectors to create}
}
\value{
a sparse matrix (of dimension n x B) containing the resampled indicators
}
\description{
Generate a (sparse) matrix of synthetic grna indicators from a vector of fitted probabilitites.
}
