% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{get_response_matrix}
\alias{get_response_matrix}
\alias{get_grna_matrix}
\alias{get_cell_covariates}
\title{Data getter functions}
\usage{
get_response_matrix(sceptre_object)

get_grna_matrix(sceptre_object)

get_cell_covariates(sceptre_object)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}
}
\value{
\code{get_response_matrix()} returns the response matrix contained within a \code{sceptre_object}.

\code{get_grna_matrix()} returns the gRNA matrix contained within a \code{sceptre_object}.

\code{get_cell_covariates()} returns the cell covariate data frame contained within a \code{sceptre_object}.
}
\description{
The data getter functions (i.e., \code{get_response_matrix()}, \code{get_grna_matrix()}, \code{get_cell_covariates()}) return of a specified data field from a \code{sceptre_object}.
}
\examples{
# 1. create a sceptre_object
library(sceptredata)
data(highmoi_example_data)
data(grna_target_data_frame_highmoi)
sceptre_object <- import_data(
  response_matrix = highmoi_example_data$response_matrix,
  grna_matrix = highmoi_example_data$grna_matrix,
  grna_target_data_frame = grna_target_data_frame_highmoi,
  moi = "high",
  extra_covariates = highmoi_example_data$extra_covariates,
  response_names = highmoi_example_data$gene_names
)

# 2. extract data fields from the sceptre_object
response_matrix <- get_response_matrix(sceptre_object)
grna_matrix <- get_grna_matrix(sceptre_object)
cell_covariates <- get_cell_covariates(sceptre_object)
}
