% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_2.R
\name{get_result}
\alias{get_result}
\title{Get result}
\usage{
get_result(sceptre_object, analysis)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}

\item{analysis}{a string indicating the name of the analysis whose results we are querying, one of \code{"run_calibration_check"}, \code{"run_power_check"}, or \code{"run_discovery_analysis"}.}
}
\value{
a data frame containing the results of the analysis
}
\description{
\code{get_result()} returns a data frame containing the result of a calibration check, power check, or discovery analysis. We pass as arguments \code{sceptre_object} and \code{analysis}, where the latter is a string indicating the function whose results we are querying. The output is a data frame, the rows of which correspond to target-response pairs, and the columns of which are as follows: \code{response_id}, \code{grna_target}, \code{n_nonzero_trt}, \code{n_nonzero_cntrl}, \code{pass_qc} (a \code{TRUE}/\code{FALSE} value indicating whether the pair passes pairwise QC), \code{p_value}, \code{log_2_fold_change}, and \code{significant} (a \code{TRUE}/\code{FALSE} value indicating whether the pair is called as significant). The p-value contained within the \code{p_value} column is a raw (i.e., non-multiplicity-adjusted) p-value. See \href{https://timothy-barry.github.io/sceptre-book/sceptre.html#sec-sceptre_write_outputs_to_directory}{Section 8 of the introductory chapter in the manual} for more information about this function.
}
\note{
If \code{output_amount} is set to \code{2} or \code{3} in \code{run_calibration_check()}, \code{run_power_check()}, or \code{run_discovery_analysis()}, then the result data frame contains additional columns; see \href{https://timothy-barry.github.io/sceptre-book/run-calibration-check.html#sec-run_calibration_check_output_amount}{Chapter 6 in the manual} for more information.
}
\examples{
library(sceptredata)
data(highmoi_example_data)
data(grna_target_data_frame_highmoi)
# import data
sceptre_object <- import_data(
  response_matrix = highmoi_example_data$response_matrix,
  grna_matrix = highmoi_example_data$grna_matrix,
  grna_target_data_frame = grna_target_data_frame_highmoi,
  moi = "high",
  extra_covariates = highmoi_example_data$extra_covariates,
  response_names = highmoi_example_data$gene_names
)
positive_control_pairs <- construct_positive_control_pairs(sceptre_object)
pc_result <- sceptre_object |>
  set_analysis_parameters(
    side = "left",
    resampling_mechanism = "permutations",
    positive_control_pairs = positive_control_pairs
  ) |>
  assign_grnas(method = "thresholding") |>
  run_qc() |>
  run_power_check() |>
  get_result("run_power_check")
}
