% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lowmoi.R
\docType{data}
\name{grna_group_data_frame_lowmoi}
\alias{grna_group_data_frame_lowmoi}
\title{gRNA group data frame}
\format{
An object of class \code{data.frame} with 110 rows and 2 columns.
}
\usage{
data(grna_group_data_frame_lowmoi)
}
\description{
A data frame containing the gRNA group information. The data frame has two columns: \code{grna_id} and \code{grna_group}. The former column provides the ID of each individual gRNA, and the latter specifies the group to which each individual gRNA belongs. gRNAs that target the same gene transcription start site are assigned to the same group. Non-targeting gRNAs are assigned to a group labeled "non-targeting".
}
\section{Overview}{

The example low MOI data in the \code{sceptre} package are taken from the paper "Characterizing the molecular regulation of inhibitory immune checkpoints with multimodal single-cell screens" by Papalexi et al., 2021. The authors used CRISPRko to target 26 genes in 20,729 THP-1 cells and measured the effects of these perturbations via single-cell sequencing. The authors designed three or four gRNAs to target each gene, as well as a library of nine non-targeting gRNAs. The example data include 290 (or 2\\%) randomly selected genes from the original dataset.
}

\keyword{datasets}
