% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_highmoi.R
\docType{data}
\name{gRNA_groups_table_highmoi}
\alias{gRNA_groups_table_highmoi}
\title{gRNA groups table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 3 columns.
}
\usage{
data(gRNA_groups_table_highmoi)
}
\description{
A data frame that maps each gRNA ID to its gRNA group and gRNA type. The columns are \code{gRNA_id} (required), \code{gRNA_group} (required), and \code{gRNA_type} (optional).
}
\section{Overview}{

This object is part of the example data. The data are taken from the paper "A genome-wide framework for mapping gene regulation via cellular genetic screens" by Gasperini et al., 2019. The authors used a CRISPRi-based assay to target 5,000+ putative enhancers in a population of K562 cells. The authors additionally targeted 200+ gene transcription start sites (TSSs) and designed a library of 50 non-targeting gRNAs to serve as negative controls. Genes, gRNAs, and cells are all down-sampled to reduce the size of the data.
}

\keyword{datasets}
