% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gRNA_grps}
\alias{gRNA_grps}
\title{gRNA groups}
\format{
An object of class \code{list} of length 5.
}
\usage{
data(gRNA_grps)
}
\description{
A named list grouping gRNAs by the chromosomal site that they target.
}
\section{Overview}{

This object is part of the example data in the \code{sceptre} package. The data are taken from the paper "Global Analysis of Enhancer Targets Reveals Convergent Enhancer-Driven Regulatory Modules" by Xie et al., 2019. The authors used a CRISPRi-based perturb-seq assay to target >500 putative enhancers in a population of K562 cells. The example data consist of a subset of the sequenced gRNAs and genes measured on 106,660 cells.
}

\keyword{datasets}
