% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grna_target_data_frame_highmoi}
\alias{grna_target_data_frame_highmoi}
\title{gRNA target data frame}
\format{
An object of class \code{data.frame} with 95 rows and 5 columns.
}
\usage{
data(grna_target_data_frame_highmoi)
}
\description{
The gRNA target data frame corresponding to the example high MOI data. A data frame containing the columns \code{grna_id} (ID of an individual gRNA), \code{grna_target} (genomic target of the gRNA), \code{chr} (target chromosome), \code{start} (start coordinate of target location), and \code{end} (end coordinate of target location).
}
\keyword{datasets}
