% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functs.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(
  response_matrix,
  grna_matrix,
  grna_target_data_frame,
  moi,
  extra_covariates = data.frame(),
  response_names = NA_character_,
  use_ondisc = FALSE,
  directory_to_write = NULL
)
}
\arguments{
\item{response_matrix}{a matrix of response UMI counts, with responses in
rows and cells in columns. The matrix should be of type \code{"matrix"},
\code{"dgCMatrix"}, \code{"dgRMatrix"}, or \code{"dgTMatrix"}. The row names of the matrix
should give the response IDs. The column names, when supplied, should give
the cell IDs.}

\item{grna_matrix}{a matrix of gRNA UMI counts, with gRNAs in rows and cells
in columns. The matrix should be of type \code{"matrix"}, \code{"dgCMatrix"},
\code{"dgRMatrix"}, or \code{"dgTMatrix"}. The row names of the matrix should give the
gRNA IDs. The column names, when supplied, should give the cell IDs.}

\item{grna_target_data_frame}{a data frame containing columns \code{grna_id} and
\code{grna_target} mapping each individual gRNA to its target. Non-targeting gRNAs
should be assigned a label of "non-targeting". Optionally,
\code{grna_target_data_frame} can contain columns \code{chr}, \code{start}, and \code{end},
giving the chromosome, start coordinate, and end coordiante, respectively, of
each gRNA. Additionally, \code{grna_target_data_frame} can contain the column
\code{vector_id} specifying the vector to which a given gRNA belongs.}

\item{moi}{a string indicating the MOI of the dataset, either "low" or
"high".}

\item{extra_covariates}{(optional) a data frame containing extra covariates
(e.g., batch, biological replicate) beyond those that \code{sceptre} can compute.
The rows should follow the same cell ordering as the columns of
\code{response_matrix} and \code{grna_matrix}. Cell IDs may be supplied as row names.
When cell IDs are supplied for at least two of these three objects, they must
be identical.}

\item{response_names}{(optional) a vector of human-readable response names;
names with the prefix "MT-" are taken as mitochondrial genes and are used to
compute the covariate \code{response_p_mito}.}

\item{use_ondisc}{(default \code{FALSE}) a logical value (i.e., \code{TRUE} or \code{FALSE})
indicating whether to create an \code{ondisc}-backed \code{sceptre_object} (\code{TRUE}) or
a standard \code{sceptre_object} (\code{FALSE}).}

\item{directory_to_write}{(optional) a file path to a directory in which to
write the backing \code{.odm} files for the response and gRNA expression matrices.
Must be supplied if \code{use_ondisc} is set to \code{TRUE}.}
}
\value{
an initialized \code{sceptre_object}
}
\description{
\code{import_data()} imports data from a collection of R objects to create a
\code{sceptre_object}. Users can create either a standard \code{sceptre} object or an
\code{ondisc}-backed \code{sceptre} object; the latter is more appropriate for
large-scale data. See
\href{https://timothy-barry.github.io/sceptre-book/import-data.html#import-data-from-a-collection-of-r-objects}{Chapter 1 of the manual}
for more detailed information about this function.
}
\examples{
data("lowmoi_example_data")
# 1. initialize a standard sceptre_object from R objects
sceptre_object <- import_data(
  response_matrix = lowmoi_example_data$response_matrix,
  grna_matrix = lowmoi_example_data$grna_matrix,
  grna_target_data_frame = lowmoi_example_data$grna_target_data_frame,
  extra_covariates = lowmoi_example_data$extra_covariates,
  moi = "low"
)

# 2. initialize an ondisc-backed sceptre_object from R objects
sceptre_object <- import_data(
  response_matrix = lowmoi_example_data$response_matrix,
  grna_matrix = lowmoi_example_data$grna_matrix,
  grna_target_data_frame = lowmoi_example_data$grna_target_data_frame,
  extra_covariates = lowmoi_example_data$extra_covariates,
  moi = "low",
  use_ondisc = TRUE,
  directory_to_write = tempdir()
)
}
