% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functs.R
\name{import_data_from_parse}
\alias{import_data_from_parse}
\title{Import data from Parse (experimental)}
\usage{
import_data_from_parse(
  gene_mat_fp,
  grna_mat_fp,
  all_genes_fp,
  all_grnas_fp,
  moi,
  grna_target_data_frame,
  extra_covariates = data.frame()
)
}
\arguments{
\item{gene_mat_fp}{file path to the gene count_matrix.mtx file}

\item{grna_mat_fp}{file path to the gRNA count_matrix.mtx file}

\item{all_genes_fp}{file path to the all_genes.csv file containing the gene IDs}

\item{all_grnas_fp}{file path to the all_guides.csv file containing the gRNA IDs}

\item{moi}{a string indicating the MOI of the dataset, either "low" or "high"}

\item{grna_target_data_frame}{a data frame containing columns \code{grna_id} and \code{grna_target} mapping each individual gRNA to its target}

\item{extra_covariates}{(optional) a data frame containing extra covariates (e.g., batch, biological replicate) beyond those that \code{sceptre} can compute}
}
\value{
an initialized \code{sceptre_object}
}
\description{
\code{import_data_from_parse()} imports data from the output of the Parse count matrix generation program. See \href{https://timothy-barry.github.io/sceptre-book/sceptre.html#sec-whole_game_import_data}{Section 1 of the introductory chapter in the manual} for more information about this function. It is assumed that the data are stored in a single set of files (as opposed to multiple sets of files corresponding to, e.g., different samples).
}
\details{
\code{import_data_from_parse()} is experimental, and the API of this function is subject to change. We expect the API to solidify as we learn more about the Parse platform and the structure of the Parse count matrix generation program output.
}
