% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{initialize_directories}
\alias{initialize_directories}
\title{Initialize directories}
\usage{
initialize_directories(storage_location)
}
\arguments{
\item{storage_location}{file path to a directory in which to store the intermediate and final results}
}
\value{
named vector containing the file paths of gene_precomp_dir, gRNA_precomp_dir, results_dir, and log_dir.
}
\description{
NOTE: DO NOT USE THIS FUNCTION DIRECTLY. INSTEAD, USE THE "AT-SCALE" BASH SCRIPT,
WHICH IS DESCRIBED HERE: https://timothy-barry.github.io/sceptre/articles/sceptre-at-scale.html
}
\details{
Initializes the offsite directory structure.
}
