% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctransform_functions.R
\name{log_geom_mean}
\alias{log_geom_mean}
\title{Log geometric mean}
\usage{
log_geom_mean(x, eps = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{eps}{small number to add to each value (default 1)}
}
\value{
the log-transformed geometric means
}
\description{
Log geometric mean
}
