% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lowmoi_example_data}
\alias{lowmoi_example_data}
\title{Example low-MOI data}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(lowmoi_example_data)
}
\description{
The example low-MOI data are a small simulated dataset modeled on that in the
paper "Characterizing the molecular regulation of inhibitory immune
checkpoints with multimodal single-cell screens" by Papalexi et al., 2021.
This dataset includes gene-targeting CRISPRko perturbations as well as non-
targeting perturbations.
}
\details{
The example data are stored in a list containing four components:
\itemize{
\item \code{response_matrix}: the gene-by-cell expression matrix
\item \code{grna_matrix}: the gRNA-by-cell expression matrix
\item \code{grna_target_data_frame}: a data frame containing the columns \code{grna_id} (ID
of an individual gRNA) and \code{grna_target} (genomic target of the gRNA)
\item \code{extra_covariates}: a data frame with a single column, \code{batch}, specifying
the batch in which each cell was sequenced (\code{batch_1} or \code{batch_2}); the row
names give the cell IDs and match the matrix column names
}
}
\keyword{datasets}
