% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{make_volcano_plot}
\alias{make_volcano_plot}
\title{Title}
\usage{
make_volcano_plot(
  discovery_result,
  p_thresh,
  x_limits = c(-1.5, 1.5),
  transparency = 0.5,
  point_size = 0.55
)
}
\arguments{
\item{point_size}{}
}
\description{
Title
}
