% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lowmoi.R
\docType{data}
\name{pc_pairs_lowmoi}
\alias{pc_pairs_lowmoi}
\title{gRNA group data frame}
\format{
An object of class \code{data.frame} with 10 rows and 2 columns.
}
\usage{
data(pc_pairs_lowmoi)
}
\description{
A data frame containing the positive control response-gRNA group pairs.
}
\section{Overview}{

The example low MOI data in the \code{sceptre} package are taken from the paper "Characterizing the molecular regulation of inhibitory immune checkpoints with multimodal single-cell screens" by Papalexi et al., 2021. The authors used CRISPRko to target 26 genes in 20,729 THP-1 cells and measured the effects of these perturbations via single-cell sequencing. The authors designed three or four gRNAs to target each gene, as well as a library of nine non-targeting gRNAs. The example data include 290 (or 2\\%) randomly selected genes from the original dataset.
}

\keyword{datasets}
