% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_covariates}
\alias{plot_covariates}
\title{Plot covariates}
\usage{
plot_covariates(
  sceptre_object,
  response_n_umis_range = c(0.01, 0.99),
  response_n_nonzero_range = c(0.01, 0.99),
  p_mito_threshold = 0.2
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{response_n_umis_range}{TBD}

\item{response_n_nonzero_range}{TBD}

\item{p_mito_threshold}{TBD}
}
\value{
TBD
}
\description{
Plot covariates
}
