% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_grna_count_distributions}
\alias{plot_grna_count_distributions}
\title{Plot the number of cells per gRNA count}
\usage{
plot_grna_count_distributions(
  sceptre_object,
  n_grnas_to_plot = 4L,
  grnas_to_plot = NULL,
  threshold = NULL
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{n_grnas_to_plot}{(optional; default \code{4}) the number of different gRNAs to plot}

\item{grnas_to_plot}{(optional; default NULL) the names of specific gRNAs to plot. If \code{NULL} then random ones are picked.}

\item{threshold}{(optional; default \code{NULL}) an integer representing a gRNA count cut-off; if provided, the bins of length 1 will go up to and include this value, after which the exponentially growing bins begin. A vertical line is also drawn at this value. If \code{NULL} then 10 is the largest gRNA count with its own bin. Non-integer values will be rounded.}
}
\value{
a single \code{ggplot2} plot. The x-axis is a piecewise linear-log scale, with bins of size 1 going from gRNA counts of 0 up to max(10, \code{threshold}), and then the bins grow exponentially fast in size.
}
\description{
Plot the number of cells per gRNA count
}
\examples{
TODO
}
