% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_resampling_distribution}
\alias{plot_resampling_distribution}
\title{Plot resampling distribution}
\usage{
plot_resampling_distribution(result_df, row_number)
}
\arguments{
\item{result_df}{the output of a call to \code{run_sceptre_lowmoi} (or \code{run_sceptre_highmoi_experimental}). NOTE: the argument \code{output_amount} should be set to \code{3} so that the null statistics are returned.}

\item{row_number}{an integer identifying the row of the results data frame to visualize}
}
\value{
a plot of the resampling distribution of null test statistics.
}
\description{
Plots the resampling distribution of null test statistics for a given response-gRNA group pair.
}
\details{
The fitted parametric curve is superimposed over the histogram in purple (if applicable). The original test statistic, meanwhile, is plotted as a vertical blue line.
}
\examples{
library(Matrix)
set.seed(5)

# 0. load the data associated with the experiment
data(response_matrix_lowmoi) # response-by-cell expression matrix
data(grna_matrix_lowmoi) # gRNA-by-cell expression matrix
data(covariate_data_frame_lowmoi) # cell-by-covariate data frame
data(grna_group_data_frame_lowmoi) # gRNA group information
response_grna_group_pairs <- generate_all_pairs(response_matrix_lowmoi,
grna_group_data_frame_lowmoi) |> dplyr::sample_n(30)
formula_object <- formula(~log(response_n_umis) + log(response_n_nonzero) +
bio_rep + p_mito)

# 1. apply sceptre
discovery_result <- run_sceptre_lowmoi(response_matrix = response_matrix_lowmoi,
grna_matrix = grna_matrix_lowmoi,
covariate_data_frame = covariate_data_frame_lowmoi,
grna_group_data_frame = grna_group_data_frame_lowmoi,
formula_object = formula_object,
response_grna_group_pairs = response_grna_group_pairs,
calibration_check = FALSE,
output_amount = 3)

# 2. visualize a resampling distribution
plot_resampling_distribution(discovery_result, 1)
}
