% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_skew_t}
\alias{plot_skew_t}
\title{Plot fitted skew-t distribution}
\usage{
plot_skew_t(sceptre_result, interval = c(-4, 4))
}
\arguments{
\item{sceptre_result}{output of \code{run_sceptre_gRNA_gene_pair} when \code{full_output} is set to TRUE}

\item{interval}{(optional) interval over which the distribution is plotted}
}
\value{
a ggplot object
}
\description{
Plots the skew-t distribution fitted to the resampled test statistics, alongside the "ground truth" test statistic derived from the raw data.
}
\details{
One also can call this function on a given row of the data frame outputted by \code{run_sceptre_in_memory} when \code{full_output} is set to TRUE.
}
\examples{
data(gene_matrix); data(gRNA_matrix); data(covariate_matrix)
gene_expressions <- gene_matrix[1,]
gRNA_expressions <- gRNA_matrix[1,]
# run method
result <- run_sceptre_gRNA_gene_pair(gene_expressions, gRNA_expressions,
covariate_matrix, "left", full_output = TRUE)
# plot result
plot_skew_t(result)
}
