% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_helpers.R
\name{print,sceptre_object-method}
\alias{print,sceptre_object-method}
\title{Print}
\usage{
\S4method{print}{sceptre_object}(x)
}
\arguments{
\item{x}{a \code{sceptre_object}}
}
\description{
\code{print()} prints information about the dataset and the status of the analysis to the console.
}
