% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{regularize_gene_sizes_at_scale}
\alias{regularize_gene_sizes_at_scale}
\title{Regularize genes at scale}
\usage{
regularize_gene_sizes_at_scale(
  gene_precomp_dir,
  regularization_amount,
  log_dir
)
}
\arguments{
\item{gene_precomp_dir}{location of the gene precomputation directory}

\item{regularization_amount}{amount of regularization to apply to thetas (>= 0)}

\item{log_dir}{location of the directory in which to sink the log file}
}
\description{
NOTE: DO NOT USE THIS FUNCTION DIRECTLY. INSTEAD, USE THE "AT-SCALE" BASH SCRIPT,
WHICH IS DESCRIBED HERE: https://timothy-barry.github.io/sceptre/articles/sceptre-at-scale.html
}
\details{
Regularizes the estimated gene sizes.
}
