% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctransform_functions.R
\name{regularize_thetas}
\alias{regularize_thetas}
\title{Regularize thetas}
\usage{
regularize_thetas(
  genes_log_gmean,
  theta,
  theta_regularization = "log_theta",
  bw_adjust = 3,
  plot_me = FALSE
)
}
\arguments{
\item{genes_log_gmean}{the log-transformed vector of the geometric mean of the gene expressions}

\item{theta}{the vector of unregularized gene dispersions}

\item{theta_regularization}{transformation to apply to the thetas before fitting nonparametric regression; options "log_theta" and "od_factor;" defaults to "log_theta"}

\item{bw_adjust}{amount of regularization (greater value -> more regularization). 0 corresponds to no regularization at all. Default value 3.}

\item{plot_me}{create informative plots? (default FALSE)}
}
\value{

}
\description{
Regularize thetas
}
\examples{
genes_log_gmean <- c(runif(1000, 2, 5))
names(genes_log_gmean) <- paste0("gene_", 1:1000)
theta <- genes_log_gmean * 0.5 + rnorm(100)
theta_regularization <- "log_theta"
bw_adjust <- 3
theta_reg <- regularize_thetas(genes_log_gmean, theta, theta_regularization, 0)
}
