% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_highmoi.R
\docType{data}
\name{response_matrix_highmoi}
\alias{response_matrix_highmoi}
\title{Gene expression matrix}
\format{
An object of class \code{dgTMatrix} with 341 rows and 40000 columns.
}
\usage{
data(response_matrix_highmoi)
}
\description{
A gene-by-cell expression matrix.
}
\section{Overview}{

This object is part of the example data. The data are taken from the paper "A genome-wide framework for mapping gene regulation via cellular genetic screens" by Gasperini et al., 2019. The authors used a CRISPRi-based assay to target 5,000+ putative enhancers in a population of K562 cells. The authors additionally targeted 200+ gene transcription start sites (TSSs) and designed a library of 50 non-targeting gRNAs to serve as negative controls. Genes, gRNAs, and cells are all down-sampled to reduce the size of the data.
}

\keyword{datasets}
