% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_constructor_functs.R
\name{return_all_pairs}
\alias{return_all_pairs}
\title{Return all pairs}
\usage{
return_all_pairs(sceptre_object, grna_groups_to_exclude = NULL)
}
\arguments{
\item{grna_groups_to_exclude}{}
}
\description{
Return all pairs
}
