% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_constructor_functs.R
\name{return_cis_pairs}
\alias{return_cis_pairs}
\title{Return cis pairs}
\usage{
return_cis_pairs(
  sceptre_object,
  grna_groups_to_exclude = character(),
  distance_threshold = 500000L,
  ref_genome = "10X_GRCh38_2020"
)
}
\arguments{
\item{distance_threshold}{}
}
\description{
Return cis pairs
}
