% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_method_functions.R
\name{run_gene_precomputation}
\alias{run_gene_precomputation}
\title{Run gene precomputation}
\usage{
run_gene_precomputation(expressions, covariate_matrix, gene_precomp_dispersion)
}
\arguments{
\item{expressions}{the vector of gene expressions}

\item{covariate_matrix}{the cell-specific covariate matrix}

\item{gene_precomp_dispersion}{the pre-computed dispersion parameter (NULL if none)}
}
\value{
a named list containing two items: offsets and dispersion.
}
\description{
This function runs the precomputation for a given gene. In particlar, it fits an NB regression of expression against covariates. The estimate of theta (i.e., the NB dispersion parameter) is obtained from glm.nb function. This is sensible as, under the null hypothesis, the NB model without the gRNA indicator is true. Offsets are obtained by log-transforming the fitted values.
}
\examples{
sim_dat <- simulate_crispr_screen_data(num_cells = 1000,
grna_mean_prob = 0.2,
covariate_sampler = list(cell_size = rnorm, cell_cycle = runif),
mRNA_mean_expression = 40,
gRNA_effect = 0,
covariate_effects = c(0.5, 1),
zero_inflation = 0,
neg_binom_size = 2)
expressions <- sim_dat$Y
covariate_matrix <- sim_dat$covariate_df
gene_precomp <- run_gene_precomputation(expressions, covariate_matrix, NULL, NULL)
}
