% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{run_gene_precomputation_at_scale}
\alias{run_gene_precomputation_at_scale}
\title{Run gene precomputation at scale (depricated)}
\usage{
run_gene_precomputation_at_scale(
  pod_id,
  gene_precomp_dir,
  cell_gene_expression_matrix,
  ordered_gene_ids,
  covariate_matrix,
  cell_subset = NULL,
  log_dir = NULL,
  gene_sizes = NULL
)
}
\arguments{
\item{pod_id}{ID of the gene pod on which to run the precomputation}

\item{gene_precomp_dir}{file path to the gene precomputation directory}

\item{cell_gene_expression_matrix}{an FBM containing the expression data (rows cells, columns genes)}

\item{ordered_gene_ids}{the gene IDs (i.e., names)}

\item{covariate_matrix}{the cell-specific covariate matrix}

\item{cell_subset}{(optional) integer vector identifying the cells to use in the model}

\item{log_dir}{file path to the directory in which to sink the log output}

\item{gene_sizes}{(optional) a vector of already-estimated (or known) gene sizes}
}
\description{
Run gene precomputation at scale (depricated)
}
\examples{
pod_id <- 1
gene_precomp_dir <- "/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/precomp/gene/"
cell_gene_expression_matrix_metadata <- readRDS("/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/processed/expression_FBM_metadata.rds")
cell_gene_expression_matrix <- load_fbm(cell_gene_expression_matrix_metadata)
covariate_matrix <- read.fst("/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/processed/cell_covariate_model_matrix.fst")
cell_subset <- readRDS("/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/processed/cells_to_keep.rds")
ordered_gene_ids <- readRDS("/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/processed/ordered_gene_ids.RDS")
}
