% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{run_gene_precomputation_at_scale_round_1}
\alias{run_gene_precomputation_at_scale_round_1}
\title{Run_gene_precomputation_at_scale_round_1}
\usage{
run_gene_precomputation_at_scale_round_1(
  pod_id,
  gene_precomp_dir,
  gene_matrix,
  covariate_matrix,
  regularization_amount,
  log_dir
)
}
\arguments{
\item{pod_id}{pod id}

\item{gene_precomp_dir}{location of the gene precomputation directory}

\item{gene_matrix}{a gene-by cell expression matrix; the rows (i.e., gene IDs) should be named}

\item{covariate_matrix}{the cell-specific matrix of technical factors, ideally containing the following covariates: log-transformed gene library size, log-transformed gRNA library size, percent mitochondrial reads, and batch}

\item{regularization_amount}{non-negative number specifying the amount of regularization to apply to the negative binomial dispersion parameter estimates}

\item{log_dir}{directory in which to sink the log file}
}
\description{
This function runs the first round of gene precomputations. In particular, it computes the raw dispersion estimate
and log geometric mean of each gene. It saves the results in the gene_precomp directory.
}
