% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{run_gene_precomputation_at_scale_round_2}
\alias{run_gene_precomputation_at_scale_round_2}
\title{Run gene precomputation at scale round 2}
\usage{
run_gene_precomputation_at_scale_round_2(
  pod_id,
  gene_precomp_dir,
  gene_matrix,
  covariate_matrix,
  regularization_amount,
  log_dir
)
}
\arguments{
\item{pod_id}{pod id}

\item{gene_precomp_dir}{gene precomp dir}

\item{gene_matrix}{the gene expression matrix, stored as an ondisc_matrix}

\item{covariate_matrix}{covariate matrix}

\item{regularization_amount}{the amount of regularization to apply to the estimated negative binomial size parameters}

\item{log_dir}{directory in which to sink the logs}
}
\description{
NOTE: DO NOT USE THIS FUNCTION DIRECTLY. INSTEAD, USE THE "AT-SCALE" BASH SCRIPT,
WHICH IS DESCRIBED HERE: https://timothy-barry.github.io/sceptre/articles/sceptre-at-scale.html
}
\details{
Runs the second round of gene precomputations.
}
