% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomp_functs_v2.R
\name{run_gene_precomputation_v2}
\alias{run_gene_precomputation_v2}
\title{Run gene precomputation (v2)}
\usage{
run_gene_precomputation_v2(expressions, covariate_matrix)
}
\arguments{
\item{expressions}{the numeric vector of gene expressions}

\item{covariate_matrix}{the covariate matrix on which to regress (NOTE: should contain an interecept term)}
}
\value{
a named vector of fitted coefficients, alongside the fitted size parameter (named "gene_theta")
}
\description{
Runs precomputation on a gene
}
