% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{run_gRNA_gene_pair_analysis_at_scale}
\alias{run_gRNA_gene_pair_analysis_at_scale}
\title{Run gRNA-gene pair analysis at scale}
\usage{
run_gRNA_gene_pair_analysis_at_scale(
  pod_id,
  gene_precomp_dir,
  gRNA_precomp_dir,
  results_dir,
  log_dir,
  expression_matrix,
  perturbation_matrix,
  covariate_matrix,
  regularization_amount,
  side,
  seed,
  B
)
}
\arguments{
\item{pod_id}{id of the pod to compute}

\item{gene_precomp_dir}{the directory containing the results of the gene precomputation}

\item{gRNA_precomp_dir}{the directory containing the results of the gRNA precomputation}

\item{results_dir}{directory in which to store the results}

\item{log_dir}{(optional) directory in which to sink the log file}

\item{covariate_matrix}{the cell-covariate matrix}

\item{regularization_amount}{amount of regularuzation to apply to thetas}

\item{side}{(optional, default left) sidedness of test}

\item{seed}{(optional) seed to pass to the randomization algorithm}

\item{B}{number of bootstrap resamples (default 500)}

\item{cell_gene_expression_matrix}{an FBM containing the cell-by-gene expressions}

\item{ordered_gene_ids}{the gene ids of the cell-gene expression matrix}

\item{gRNA_indicator_matrix_fp}{a file-path to the gRNA indicator matrix (assumed to be stored as a .fst file)}

\item{cell_subset}{(optional) an integer vector containing the cells to subset}
}
\description{
Runs the gene-gRRNA pair ananysis across an entire pod of gene-gRNA pairs.
}
