% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomp_functs_v2.R
\name{run_grna_precomputation_v2}
\alias{run_grna_precomputation_v2}
\title{Run grna precomputation (v2)}
\usage{
run_grna_precomputation_v2(indicators, covariate_matrix)
}
\arguments{
\item{indicators}{the binary vector of grna indicators}

\item{covariate_matrix}{the covariate matrix on which to regress (NOTE: should contain an interecept term)}
}
\value{
a named vector of fitted coefficients
}
\description{
Runs precomputation on a vector of grna indicators
}
