% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm_functions.R
\name{run_precomputation_gcm}
\alias{run_precomputation_gcm}
\title{Run GCM precomputation}
\usage{
run_precomputation_gcm(expressions, covariate_matrix, family)
}
\arguments{
\item{expressions}{a numeric (or integer) vector of gene expressions or gRNA indicators}

\item{covariate_matrix}{a matrix of technical factors on which to regress the expressions or indicators}

\item{a}{string giving the family to use in the regression (one of "poisson" and "binomial")}
}
\value{
a list with element "resids", the vector of residuals
}
\description{
Runs the precomputation for the "GCM" mode of inference
}
