% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_1.R
\name{run_qc}
\alias{run_qc}
\title{Run QC}
\usage{
run_qc(
  sceptre_object,
  n_nonzero_trt_thresh = 7L,
  n_nonzero_cntrl_thresh = 7L,
  response_n_umis_range = c(0.01, 0.99),
  response_n_nonzero_range = c(0.01, 0.99),
  p_mito_threshold = 0.2,
  additional_cells_to_remove = integer()
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{n_nonzero_trt_thresh}{TBD}

\item{n_nonzero_cntrl_thresh}{TBD}

\item{response_n_umis_range}{TBD}

\item{response_n_nonzero_range}{TBD}

\item{p_mito_threshold}{TBD}

\item{additional_cells_to_remove}{TBD}
}
\description{
Run QC
}
