% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_interface_v2.R
\name{run_sceptre}
\alias{run_sceptre}
\title{Run \code{sceptre}}
\usage{
run_sceptre(
  response_matrix,
  grna_matrix,
  covariate_data_frame,
  grna_group_data_frame,
  response_grna_group_pairs,
  moi,
  formula_object,
  calibration_check,
  control_group,
  resampling_mechanism,
  n_nonzero_trt_thresh = 7L,
  n_nonzero_cntrl_thresh = 7L,
  grna_assign_threshold = 5L,
  return_debugging_metrics = FALSE,
  return_resampling_dist = FALSE,
  fit_skew_normal = TRUE,
  calibration_group_size = NULL,
  n_calibration_pairs = NULL,
  B1 = 499L,
  B2 = 4999L,
  B3 = 24999L,
  print_progress = TRUE
)
}
\arguments{
\item{print_progress}{}
}
\description{
Run \code{sceptre}
}
\examples{
\dontrun{

library(Matrix)

#################
# LOW MOI example
#################

# 0. load the data associated with the experiment
data(response_matrix_lowmoi) # response-by-cell expression matrix
data(grna_matrix_lowmoi) # gRNA-by-cell expression matrix
data(covariate_data_frame_lowmoi) # cell-by-covariate data frame
data(grna_group_data_frame_lowmoi) # gRNA group information

# 1. obtain the set of pairs to analyze
response_grna_group_pairs <- generate_all_pairs(response_matrix_lowmoi,
grna_group_data_frame_lowmoi)

# 2. set the formula object
formula_object <- formula(~log(response_n_umis) + log(response_n_nonzero) +
bio_rep + p_mito)

# 3. run the calibration check analysis (NOTE: calibration_check set to TRUE)
calibration_result <- run_sceptre(response_matrix = response_matrix_lowmoi,
grna_matrix = grna_matrix_lowmoi,
covariate_data_frame = covariate_data_frame_lowmoi,
grna_group_data_frame = grna_group_data_frame_lowmoi,
formula_object = formula_object,
response_grna_group_pairs = response_grna_group_pairs,
calibration_check = TRUE)

# 4. plot the calibration result to ensure adequate calibration
plot_calibration_result(calibration_result)

# 5. run the discovery analysis (NOTE: calibration_check set to FALSE)
discovery_result <- run_sceptre(response_matrix = response_matrix_lowmoi,
grna_matrix = grna_matrix_lowmoi,
covariate_data_frame = covariate_data_frame_lowmoi,
grna_group_data_frame = grna_group_data_frame_lowmoi,
formula_object = formula_object,
response_grna_group_pairs = response_grna_group_pairs,
calibration_check = FALSE)

# 6. compare discovery p-values to the negative control p-values; make a volcano plot
compare_calibration_and_discovery_results(calibration_result, discovery_result)
make_volcano_plot(discovery_result)

# 7. obtain the discovery set for downstream analysis
discovery_set <- obtain_discovery_set(discovery_result)

##################
# HIGH MOI example
##################
data(response_matrix_highmoi) # response-by-cell expression matrix
data(grna_matrix_highmoi) # gRNA-by-cell expression matrix
data(covariate_data_frame_highmoi) # cell-by-covariate data frame
data(grna_group_data_frame_highmoi) # gRNA group information
data(discovery_pairs_highmoi)

# 1. set the pairs to analyze
response_grna_group_pairs <- discovery_pairs_highmoi

# 2. set the formula object
formula_object <- formula(~log(grna_n_umis) + log(gene_n_umis) + batch + p_mito)

# 3. automatic calibration check is not yet programmed in high MOI

# 4. run discovery analysis
discovery_result <- run_sceptre(response_matrix = response_matrix_highmoi,
grna_matrix = grna_matrix_highmoi,
covariate_data_frame = covariate_data_frame_highmoi,
grna_group_data_frame = grna_group_data_frame_highmoi,
response_grna_group_pairs = discovery_pairs_highmoi,
moi = "high",
formula_object = formula_object,
calibration_check = FALSE,
control_group = "complement",
resampling_mechanism = "permutations")
}
}
