% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_interface.R
\name{run_sceptre_highmoi}
\alias{run_sceptre_highmoi}
\title{Title}
\usage{
run_sceptre_highmoi(
  response_matrix,
  grna_matrix,
  covariate_data_frame,
  grna_group_data_frame,
  formula_object,
  response_grna_group_pairs,
  calibration_check,
  threshold = 5,
  n_nonzero_trt_thresh = 7L,
  n_nonzero_cntrl_thresh = 7L,
  discovery_test_stat = "exact",
  return_debugging_metrics = FALSE,
  return_resampling_dist = FALSE,
  fit_skew_normal = TRUE,
  calibration_group_size = NULL,
  n_calibration_pairs = NULL,
  B1 = 499L,
  B2 = 4999L,
  B3 = 24999L,
  regression_method = "poisson_glm",
  print_progress = TRUE
)
}
\arguments{
\item{print_progress}{}
}
\description{
Title
}
\examples{
data(covariate_matrix_highmoi)
data(gene_matrix_highmoi)
data(gRNA_matrix_highmoi)
data(gRNA_groups_table_highmoi)

# modify the grna group table to get it into the "standard" format
gRNA_groups_table_highmoi$gRNA_group[gRNA_groups_table_highmoi$gRNA_type == "non_target"] <- "non-targeting"
gRNA_groups_table_highmoi <- gRNA_groups_table_highmoi |>
dplyr::rename("grna_id" = "gRNA_id", "grna_group" = "gRNA_group") |>
dplyr::mutate(gRNA_type = NULL)

response_matrix <- gene_matrix_highmoi
grna_matrix <- gRNA_matrix_highmoi
covariate_data_frame <- covariate_matrix_highmoi
grna_group_data_frame <- gRNA_groups_table_highmoi
formula_object <- formula(~ lg_gRNA_lib_size + lg_gene_lib_size + p_mito + batch)
response_grna_group_pairs <- generate_all_pairs(response_matrix, grna_group_data_frame)
calibration_check <- FALSE
}
