% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_interface_v2.R
\name{run_sceptre_highmoi_experimental}
\alias{run_sceptre_highmoi_experimental}
\title{Run \code{sceptre} (high MOI; experimental)}
\usage{
run_sceptre_highmoi_experimental(
  response_matrix,
  grna_matrix,
  covariate_data_frame,
  grna_group_data_frame,
  formula_object,
  calibration_check,
  response_grna_group_pairs = NULL,
  resampling_mechanism = "default",
  n_nonzero_trt_thresh = 7L,
  n_nonzero_cntrl_thresh = 7L,
  side = "both",
  grna_assign_threshold = 5L,
  output_amount = 1L,
  fit_parametric_curve = TRUE,
  calibration_group_size = NULL,
  n_calibration_pairs = NULL,
  B1 = 499L,
  B2 = 4999L,
  B3 = 24999L,
  print_progress = TRUE
)
}
\description{
Run \code{sceptre} (high MOI; experimental)
}
\examples{
\dontrun{
library(Matrix)

# 0. load the data
data(response_matrix_highmoi_experimental) # response-by-cell expression matrix
data(grna_matrix_highmoi_experimental) # gRNA-by-cell expression matrix
data(covariate_data_frame_highmoi_experimental) # cell-by-covariate data frame
data(grna_group_data_frame_highmoi_experimental) # gRNA group information
data(discovery_pairs_highmoi_experimental)

# 1. set the formula object
formula_object <- formula(~log(grna_n_umis) + log(gene_n_umis) + batch + p_mito)

# 2. run the calibration check
calibration_result <- run_sceptre_highmoi_experimental(
response_matrix = response_matrix_highmoi_experimental,
grna_matrix = grna_matrix_highmoi_experimental,
covariate_data_frame = covariate_data_frame_highmoi_experimental,
grna_group_data_frame = grna_group_data_frame_highmoi_experimental,
response_grna_group_pairs = discovery_pairs_highmoi_experimental,
formula_object = formula_object,
side = "left",
calibration_check = TRUE)

# 3. verify calibration
plot_calibration_result(calibration_result)

# 4. run discovery analysis
discovery_result <- run_sceptre_highmoi_experimental(
response_matrix = response_matrix_highmoi_experimental,
grna_matrix = grna_matrix_highmoi_experimental,
covariate_data_frame = covariate_data_frame_highmoi_experimental,
grna_group_data_frame = grna_group_data_frame_highmoi_experimental,
response_grna_group_pairs = discovery_pairs_highmoi_experimental,
formula_object = formula_object,
side = "left",
calibration_check = FALSE)

# 5. compare discovery p-values to the negative control p-values; make a volcano plot
compare_calibration_and_discovery_results(calibration_result, discovery_result)
make_volcano_plot(discovery_result)

# 6. obtain the discovery set
discovery_set <- obtain_discovery_set(discovery_result)

# 7. optional: positive control analysis
pc_result <- run_sceptre_highmoi_experimental(
response_matrix = response_matrix_highmoi_experimental,
grna_matrix = grna_matrix_highmoi_experimental,
covariate_data_frame = covariate_data_frame_highmoi_experimental,
grna_group_data_frame = grna_group_data_frame_highmoi_experimental,
response_grna_group_pairs = pc_pairs_highmoi_experimental,
formula_object = formula_object,
side = "left",
calibration_check = FALSE)
}
}
