% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm_functions.R
\name{run_sceptre_using_gcm}
\alias{run_sceptre_using_gcm}
\title{Run GCM inference}
\usage{
run_sceptre_using_gcm(gene_resids, gRNA_resids, side)
}
\arguments{
\item{gene_resids}{the residuals of the gene regression}

\item{gRNA_resids}{the residuals of the gRNA regression}

\item{side}{sidedness of the test; one of "left", "right", or "both"}
}
\value{
a data frame containing columns "p_value" and "z_value"
}
\description{
Run GCM inference using the gene and gRNA precomputations
}
