% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_method_functions.R
\name{run_sceptre_using_precomp}
\alias{run_sceptre_using_precomp}
\title{Run sceptre using precomputations for gRNAs and genes.}
\usage{
run_sceptre_using_precomp(
  expressions,
  gRNA_indicators,
  gRNA_precomp,
  gene_precomp_size,
  gene_precomp_offsets,
  B,
  seed
)
}
\arguments{
\item{expressions}{a vector of gene expressions (in UMI counts)}

\item{gRNA_indicators}{a vector of gRNA indicators}

\item{gRNA_precomp}{a vector of conditional probabilities for gRNA assignments}

\item{gene_precomp_size}{the pre-computed size}

\item{gene_precomp_offsets}{the pre-computed distillation offsets}

\item{B}{the number of resamples to make (default 500)}

\item{seed}{an arguement to set.seed; if null, no seed is set}
}
\value{
a p-value of the null hypothesis of no gRNA effect on gene expression
}
\description{
This function is the workhorse function of the sceptre package. It runs a distilled CRT using a negative binomial test statistic based on an expression vector, a gRNA indicator vector, an offset vector (from the distillation step), gRNA conditional probabilities, an estimate of the negative binomial size parameter, and the number of resampling replicates.
}
\details{
This currently is a one-tailed, left-sided test. Thus, it is suitable for up-regulatory elements like enhancers and promoters but not down-regulatory elements like silencers.
}
