% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceptre.R
\docType{package}
\name{sceptre}
\alias{sceptre}
\title{sceptre}
\description{
\code{sceptre} is a statistically principled, fast, memory-light, and user-friendly software for single-cell CRISPR screen analysis. sceptre achieves state-of-the-art calibration and power on single-cell CRISPR screen data by leveraging several methodological and algorithmic advances in assumption-lean and computationally efficient differential expression analysis. \code{sceptre} includes modules for the analysis of both low MOI and high MOI data.
}
