% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_crispr_screen_data}
\alias{simulate_crispr_screen_data}
\title{Simulate CRISPR screen data.}
\usage{
simulate_crispr_screen_data(
  num_cells,
  grna_mean_prob,
  covariate_sampler,
  mRNA_mean_expression,
  gRNA_effect,
  covariate_effects,
  zero_inflation,
  neg_binom_size
)
}
\arguments{
\item{num_cells}{number of cells in experiment}

\item{grna_mean_prob}{mean number of cells perturbed (absent covariates)}

\item{covariate_sampler}{a list of functions. Each function in this list takes as an argument num_cells and returns num_cells random variates.}

\item{mRNA_mean_expression}{mean mRNA expression of the cell (absent covariates)}

\item{gRNA_effect}{effect of the gRNA perturbation on mRNA expression}

\item{covariate_effects}{a numeric vector indicating the covariate effects on mRNA expression}

\item{zero_inflation}{a numeric scalar between 0 and 1 (inclusive) giving the mean fraction of cell expressions set to zero}

\item{neg_binom_size}{size parameter for the negative binomial model}
}
\value{
a list containing (i) the expression vector Y, (ii) the data frame of covariates, and (iii) the gRNA presence indicator vector.
}
\description{
A function to simulate pooled CRISPR screen data.
}
\examples{
num_cells <- 1000
grna_mean_prob <- 0.2
covariate_sampler <- list(
 cell_size = runif,
 cell_cycle = function(x) {runif(n = x, min = 0, max = 1)}
)
mRNA_mean_expression <- 40
gRNA_effect <- 4
covariate_effects <- c(2, 1)
zero_inflation <- 0
neg_binom_size <- 2
simulate_crispr_screen_data(num_cells, grna_mean_prob, covariate_sampler, mRNA_mean_expression, gRNA_effect, covariate_effects, zero_inflation, neg_binom_size)
}
