% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ondisc_functs.R
\name{write_ondisc_backed_sceptre_object}
\alias{write_ondisc_backed_sceptre_object}
\title{Write ondisc-backed sceptre object}
\usage{
write_ondisc_backed_sceptre_object(sceptre_object, sceptre_object_fp)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}

\item{sceptre_object_fp}{the file in which to store the \code{sceptre_object}}
}
\description{
Write an ondisc-backed sceptre object to disk.
}
