% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_1.R
\name{assign_grnas}
\alias{assign_grnas}
\title{Assign gRNAs to cells}
\usage{
assign_grnas(
  sceptre_object,
  method = "default",
  print_progress = TRUE,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}

\item{method}{(optional) a string indicating the method to use to assign the gRNAs to cells, one of "mixture", "thresholding", or "maximum"}

\item{print_progress}{(optional) a logical indicating whether to print progress updates}

\item{parallel}{(optional) a logical indicating whether to run the function in parallel}

\item{...}{optional method-specific additional arguments}
}
\value{
an updated \code{sceptre_object} in which the gRNA assignments have been carried out
}
\description{
\code{assign_grnas()} performs the gRNA-to-cell assignments. See \href{https://timothy-barry.github.io/sceptre-book/assign-grnas.html}{Chapter 3 of the manual} for detailed information about this function.
}
\examples{
# see example via ?sceptre
}
