% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_1.R
\name{assign_grnas}
\alias{assign_grnas}
\title{Assign gRNAs to cells}
\usage{
assign_grnas(
  sceptre_object,
  method = "default",
  hyperparameters = "default",
  print_progress = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{method}{TBD}

\item{hyperparameters}{TBD}

\item{print_progress}{TBD}

\item{parallel}{TBD}
}
\description{
Assign gRNAs to cells
}
