% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{collect_results}
\alias{collect_results}
\title{Collect results}
\usage{
collect_results(results_dir, gene_gRNA_group_pairs)
}
\arguments{
\item{results_dir}{the directory containing the results}

\item{gene_gRNA_group_pairs}{the gene-gRNA groups data frame}
}
\value{
the collated results data frame.
}
\description{
Collates the individual results files into a single result file.
}
