% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{collect_results}
\alias{collect_results}
\title{Collect results}
\usage{
collect_results(results_dir)
}
\arguments{
\item{results_dir}{the directory containing the results}
}
\value{
the collated results data frame.
}
\description{
NOTE: DO NOT USE THIS FUNCTION DIRECTLY. INSTEAD, USE THE "AT-SCALE" BASH SCRIPT,
WHICH IS DESCRIBED HERE: https://timothy-barry.github.io/sceptre/articles/sceptre-at-scale.html
}
\details{
Collates the individual results files into a single result file.
}
