% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{collect_results}
\alias{collect_results}
\title{Collect results}
\usage{
collect_results(results_dir)
}
\arguments{
\item{results_dir}{the directory containing the results}

\item{save_to_disk}{(boolean) save the collated results to the results directory? (default true)}
}
\value{
the collated results data frame.
}
\description{
Collates the individual results files into a single result file. An additional column, adjusted p-value (obtained through BH), is added to the data frame.
}
