% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functs.R
\name{combine_gRNAs}
\alias{combine_gRNAs}
\title{Combine gRNAs}
\usage{
combine_gRNAs(gRNA_matrix, gRNA_grps)
}
\arguments{
\item{gRNA_matrix}{XX}

\item{gRNA_grps}{XX}
}
\value{

}
\description{
Combine gRNAs
}
\examples{
library(magrittr)
data(gRNA_matrix)
indiv_gRNAs <- sort(row.names(gRNA_matrix))
grp_size <- 5L
idxs <- seq(from = 1, to = length(indiv_gRNAs) + grp_size, by = grp_size)
group_idxs <- seq(1, length(idxs) - 1)
gRNA_grps <- lapply(X = group_idxs, FUN = function(i) {
 gRNA_grp <- indiv_gRNAs[seq(idxs[i], idxs[i + 1] - 1)]
}) \%>\% set_names(., paste0("gRNA_grp_",  group_idxs))
combined_gRNA_matrix <- combine_gRNAs(gRNA_matrix, gRNA_grps)
}
