% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functs.R
\name{combine_gRNAs}
\alias{combine_gRNAs}
\title{Combine gRNAs}
\usage{
combine_gRNAs(gRNA_matrix, gRNA_grps)
}
\arguments{
\item{gRNA_matrix}{a gRNA-by-cell expression matrix; the matrix should have row names giving the gRNA IDs}

\item{gRNA_grps}{a named list; each entry of the list should be a character vector giving the IDs of gRNAs that target the same site, with the name of that entry being the name of the targeted site}
}
\value{
a "combined" gRNA-by-cell expression matrix, where gRNAs that target the same site have been collapsed into a single row (via addition of the constituent gRNAs)
}
\description{
Combines gRNAs that target the same site into a single "combined" gRNA via addition of the expression levels.
}
\details{
It typically is a good idea to combine gRNAs that target the same site into a single site-specific "combined" gRNA to increase statistical power. This is especially the case when gRNA expression levels are low or there are not a lot of (e.g., < 100,000) cells.
}
\examples{

# 1. First example
# `gRNA_grps` indicates that the gRNAs "AGAAGAGTCAGCCTTGCAC,"
# "GAAGAGTCAGCCTTGCACT," "GTTTAGGGAACCCAGTGCA," "GTAACTTCATTTGCAGCAA,"
# "TTACTTTTTATCAAGCCAA," "TCTAATTTAAGACCTGGGT," "AAGGTATCATTTTCTTGTC,"
# "TTAGCCATAGAACCACTAC," "TTTTCCAGTAGTGGTTCTA," and "TACATGGCATAGATCTTTA"
# all target the site "chr8:128428069-128428469". The function  collapses
# the above gRNAs into a single "combined" gRNA called
# "chr8:128428069-128428469" (and similarly for the other sites).
library(magrittr)
data(gRNA_matrix)
data(gRNA_grps)
combined_gRNA_matrix <- combine_gRNAs(gRNA_matrix, gRNA_grps)

# 2. Second example
# Here we group only the gRNAs in the first two gRNA groups;
# all other gRNAs remain ungrouped.
gRNA_grps <- gRNA_grps[1:2]
combined_gRNA_matrix_2 <- combine_gRNAs(gRNA_matrix, gRNA_grps)
}
