% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{compare_calibration_and_discovery_results}
\alias{compare_calibration_and_discovery_results}
\title{Compare calibration and discovery results}
\usage{
compare_calibration_and_discovery_results(
  calibration_result,
  discovery_result,
  alpha = 0.1,
  multiple_testing_correction = "BH"
)
}
\arguments{
\item{calibration_result}{(required) the output of a call to \code{run_sceptre_lowmoi} with \code{calibration_check} set to \code{TRUE}}

\item{discovery_result}{(required) the output of a call to \code{run_sceptre_lowmoi} with \code{calibration_check} set to \code{FALSE}}

\item{alpha}{(optional; default 0.1) the nominal type-I error control level}

\item{multiple_testing_correction}{(optional; default "BH") the multiple testing correction method (one of "BH", "bonferroni", "BY", or "holm")}
}
\value{
A ggplot2 object containing the Q-Q plot.
}
\description{
The \code{compare_calibration_and_discovery_results()} function compares two sets of p-values: the negative control p-values (i.e., those obtained from the calibration check) and the discovery p-values (i.e., those obtained from the discovery analysis). These two sets of p-values are plotted on the same Q-Q plot, with the negative control p-values colored in red and the discovery p-values colored in blue. The negative control p-values should lie along the diagonal line and fall mostly within the gray confidence band, indicating control of the false discovery rate. By contrast, the discovery p-values should lie above the diagonal line, indicating the presence of signal in the discovery set.
}
\details{
The horizontal dashed line indicates the multiple testing threshold. Blue points above the horizontal dashed line are called as discoveries.
}
\examples{
# See the example in the run_sceptre_lowmoi help file.
?run_sceptre_lowmoi
}
