% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_constructor_functs.R
\name{construct_cis_pairs}
\alias{construct_cis_pairs}
\title{Construct cis pairs}
\usage{
construct_cis_pairs(
  sceptre_object,
  distance_threshold = 500000L,
  grna_groups_to_exclude = character(),
  response_grna_group_pairs_to_exclude = data.frame(),
  ref_genome = "10X_GRCh38_2020"
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{distance_threshold}{TBD}

\item{grna_groups_to_exclude}{TBD}

\item{response_grna_group_pairs_to_exclude}{TBD}

\item{ref_genome}{TBD}
}
\value{
TBD
}
\description{
Construct cis pairs
}
