% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_constructor_functs.R
\name{construct_cis_pairs}
\alias{construct_cis_pairs}
\title{Construct cis pairs}
\usage{
construct_cis_pairs(
  sceptre_object,
  positive_control_pairs = data.frame(),
  distance_threshold = 500000L,
  exclude_positive_control_grna_targets = TRUE,
  ref_genome = "10X_GRCh38_2020"
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{distance_threshold}{TBD}

\item{ref_genome}{TBD}

\item{grna_targets_to_exclude}{TBD}

\item{response_grna_target_pairs_to_exclude}{TBD}
}
\value{
TBD
}
\description{
Construct cis pairs
}
