% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_constructor_functs.R
\name{construct_positive_control_pairs}
\alias{construct_positive_control_pairs}
\title{Construct positive control pairs}
\usage{
construct_positive_control_pairs(sceptre_object)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}
}
\value{
a data frame with columns \code{grna_target} and \code{response_id} containing the positive control pairs
}
\description{
\code{construct_positive_control_pairs()} is a helper function to facilitate construction of the positive control pairs. Positive control pairs are target-response pairs for which we know (or have strong reason to believe) that there is a regulatory relationship between the target and the response. We can use positive control pairs to verify that \code{sceptre} is sensitive (i.e., capable of detecting true associations) on the dataset under analysis. \code{construct_positive_control_pairs()} takes as an argument a \code{sceptre_object} and returns a data frame with columns \code{grna_target} and \code{response_id}, where gRNA targets and response IDs with matching names are paired. Typically, the positive control set consists of transcription start sites paired to the gene regulated by those transcription start sites. See \href{https://timothy-barry.github.io/sceptre-book/set-analysis-parameters.html#positive-control-pairs}{Section 2.2 in the manual} for more detailed information about this function.
}
\examples{
data(highmoi_example_data)
data(grna_target_data_frame_highmoi)
# import data
sceptre_object <- import_data(
  response_matrix = highmoi_example_data$response_matrix,
  grna_matrix = highmoi_example_data$grna_matrix,
  grna_target_data_frame = grna_target_data_frame_highmoi,
  moi = "high",
  extra_covariates = highmoi_example_data$extra_covariates,
  response_names = highmoi_example_data$gene_names
)
positive_control_pairs <- construct_positive_control_pairs(sceptre_object)
}
