% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_constructor_functs.R
\name{construct_positive_control_pairs}
\alias{construct_positive_control_pairs}
\title{Construct positive control pairs}
\usage{
construct_positive_control_pairs(sceptre_object)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}
}
\value{
a data frame with columns \code{grna_target} and \code{response_id} containing the positive control pairs
}
\description{
\code{construct_positive_control_pairs()} is a helper function to facilitate construction of the positive control pairs. See \href{https://timothy-barry.github.io/sceptre-book/set-analysis-parameters.html#positive-control-pairs}{Section 2.2 in the manual} for more detailed information about this function.
}
\examples{
# see example via ?sceptre
}
