% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_constructor_functs.R
\name{construct_positive_control_pairs}
\alias{construct_positive_control_pairs}
\title{Construct positive control pairs}
\usage{
construct_positive_control_pairs(sceptre_object)
}
\arguments{
\item{sceptre_object}{TBD}
}
\value{
TBD
}
\description{
Helper function to construct the positive control pairs.
}
