% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_constructor_functs.R
\name{construct_trans_pairs}
\alias{construct_trans_pairs}
\title{Construct \emph{trans} pairs}
\usage{
construct_trans_pairs(
  sceptre_object,
  positive_control_pairs = data.frame(),
  pairs_to_exclude = "none"
)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}

\item{positive_control_pairs}{(optional) the set of positive control pairs}

\item{pairs_to_exclude}{(optional; default \code{"none"}) a string specifying pairs to exclude from the \emph{trans} pairs, one of \code{"none"}, \code{"pc_pairs"}, or \code{"pairs_containing_pc_targets"}}
}
\value{
a data frame with columns \code{grna_target} and \code{response_id} containing the \emph{trans} discovery set
}
\description{
\code{construct_trans_pairs()} is a helper function to facilitate construction the set of \emph{trans} pairs. See \href{https://timothy-barry.github.io/sceptre-book/set-analysis-parameters.html#sec-set-analysis-parameters_construct_trans_pairs}{Section 2.2.2} of the manual for more detailed information about this function.
}
\details{
Typically, in screens of genes (resp., noncoding regulatory elements), we set \code{pairs_to_exclude} to \code{"pc_pairs"} (resp., \code{"pairs_containing_pc_targets"}).
}
\examples{
library(sceptredata)
# 1. low-moi, gene-targeting screen
data("lowmoi_example_data")
sceptre_object <- import_data(
  response_matrix = lowmoi_example_data$response_matrix,
  grna_matrix = lowmoi_example_data$grna_matrix,
  extra_covariates = lowmoi_example_data$extra_covariates,
  grna_target_data_frame = lowmoi_example_data$grna_target_data_frame,
  moi = "low")
positive_control_pairs <- construct_positive_control_pairs(sceptre_object)
discovery_pairs <- construct_trans_pairs(sceptre_object = sceptre_object,
                                         positive_control_pairs = positive_control_pairs,
                                         pairs_to_exclude = "pc_pairs")

# 2. high-moi, enhancer-targeting screen
data(highmoi_example_data)
data(grna_target_data_frame_highmoi)
# import data
sceptre_object <- import_data(
 response_matrix = highmoi_example_data$response_matrix,
 grna_matrix = highmoi_example_data$grna_matrix,
 grna_target_data_frame = grna_target_data_frame_highmoi,
 moi = "high",
 extra_covariates = highmoi_example_data$extra_covariates,
 response_names = highmoi_example_data$gene_names
)
positive_control_pairs <- construct_positive_control_pairs(sceptre_object)
discovery_pairs <- construct_trans_pairs(sceptre_object = sceptre_object,
                                        positive_control_pairs = positive_control_pairs,
                                        pairs_to_exclude = "pairs_containing_pc_targets")
}
