% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_constructor_functs.R
\name{construct_trans_pairs}
\alias{construct_trans_pairs}
\title{Construct trans pairs}
\usage{
construct_trans_pairs(
  sceptre_object,
  positive_control_pairs = data.frame(),
  exclude_positive_control_pairs = TRUE
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{grna_targets_to_exclude}{TBD}

\item{response_grna_target_pairs_to_exclude}{TBD}
}
\value{
TBD
}
\description{
A helper function to construct the trans pairs
}
