% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_constructor_functs.R
\name{construct_trans_pairs}
\alias{construct_trans_pairs}
\title{Construct trans pairs}
\usage{
construct_trans_pairs(
  sceptre_object,
  positive_control_pairs = data.frame(),
  exclude_positive_control_pairs = TRUE,
  exclude_positive_control_targets = FALSE
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{positive_control_pairs}{TBD}

\item{exclude_positive_control_pairs}{TBD}
}
\value{
TBD
}
\description{
A helper function to construct the trans pairs
}
