% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_and_functs.R
\name{create_sceptre_object}
\alias{create_sceptre_object}
\title{Create a \code{sceptre} object}
\usage{
create_sceptre_object(
  response_matrix,
  grna_matrix,
  covariate_data_frame,
  grna_group_data_frame,
  moi
)
}
\arguments{
\item{response_matrix}{(required) a matrix of raw expression counts. The responses (e.g., genes or proteins) should be in the rows, and the cells should be in the columns. The row names should be the unique IDs of the responses. The matrix can be a standard (dense) matrix or a sparse matrix of class \code{dgCMatrix}, \code{dgRMatrix}, or \code{dgTMatrix}.}

\item{grna_matrix}{(required) a matrix of gRNA expression counts. The gRNAs should be in the rows, and the cells should be in the columns. The row names should be the unique IDs of the gRNAs. The matrix can be a standard (dense) matrix or a sparse matrix of class \code{dgCMatrix}, \code{dgRMatrix}, or \code{dgTMatrix}. (See "Notes" for details about passing a gRNA matrix containing user-specified gRNA-to-cell assignments.)}

\item{covariate_data_frame}{(required) a data frame containing the cell-specific covariates (e.g., sequencing batch, total UMI count, etc.), with cells in the rows and covariates in the columns.}

\item{grna_group_data_frame}{(required) a data frame specifying the "group" to which each individual gRNA belongs. This data frame should contain columns \code{grna_id} and \code{grna_group}, with \code{grna_group} specifying the group membership of a given gRNA in the \code{grna_id} column. Typically, gRNAs that target the same site are grouped into the same gRNA group. Non-targeting gRNAs (if present) should be assigned a gRNA group label of "non-targeting".}

\item{moi}{the MOI of the dataset, either "low" or "high".}
}
\value{
an object of class \code{sceptre_object}
}
\description{
Create a \code{sceptre} object
}
\examples{
#################
# Low MOI example
#################
# 0. obtain the data required for a single-cell screen analysis
data(response_matrix_lowmoi) # response-by-cell expression matrix
data(grna_matrix_lowmoi) # gRNA-by-cell expression matrix
data(covariate_data_frame_lowmoi) # cell-by-covariate data frame
data(grna_group_data_frame_lowmoi) # gRNA group information

# 1. create the sceptre object
sceptre_object <- create_sceptre_object(
response_matrix = response_matrix_lowmoi,
grna_matrix = grna_matrix_lowmoi,
covariate_data_frame = covariate_data_frame_lowmoi,
grna_group_data_frame = grna_group_data_frame_lowmoi,
moi = "low")

# 2. obtain the formula object and pairs to analyze
formula_object <- formula(~log(response_n_umis) + log(response_n_nonzero) + bio_rep + p_mito)
discovery_pairs <- generate_all_pairs(sceptre_object)

# 3. prepare the analysis
sceptre_object <- prepare_analysis(
sceptre_object = sceptre_object,
formula_object = formula_object,
discovery_pairs = discovery_pairs)

# 4. run the calibration check
sceptre_object <- run_calibration_check(sceptre_object)
plot(sceptre_object)

# 5. (optional) run the power check
sceptre_object <- run_power_check(sceptre_object)
plot(sceptre_object)

# 6. run discovery analysis
sceptre_object <- run_discovery_analysis(sceptre_object)
plot(sceptre_object)

##################
# High MOI example
##################
# 0. obtain the data required for a single-cell screen analysis
data(response_matrix_highmoi_experimental)
data(grna_matrix_highmoi_experimental)
data(covariate_data_frame_highmoi_experimental)
data(grna_group_data_frame_highmoi_experimental)

# 1. create the sceptre object
sceptre_object <- create_sceptre_object(
response_matrix = response_matrix_highmoi_experimental,
grna_matrix = grna_matrix_highmoi_experimental,
covariate_data_frame = covariate_data_frame_highmoi_experimental,
grna_group_data_frame = grna_group_data_frame_highmoi_experimental,
moi = "high")

# 2. set the discovery pairs and positive control pairs
data(discovery_pairs_highmoi_experimental)
data(pc_pairs_highmoi_experimental)

# 3. prepare the analysis
sceptre_object <- prepare_analysis(
sceptre_object = sceptre_object,
discovery_pairs = discovery_pairs_highmoi_experimental,
positive_control_pairs = pc_pairs_highmoi_experimental,
resampling_mechanism = "permutations",
side = "left")

# 4. run the calibration check; plot the result
sceptre_object <- run_calibration_check(sceptre_object)
plot(sceptre_object)

# 5. (optional) run the power check
sceptre_object <- run_power_check(sceptre_object)

# 6. run discovery analysis
sceptre_object <- run_discovery_analysis(sceptre_object)
}
