% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes_and_functs.R
\name{create_sceptre_object}
\alias{create_sceptre_object}
\title{Create a \code{sceptre} object}
\usage{
create_sceptre_object(
  response_matrix,
  grna_matrix,
  covariate_data_frame,
  grna_group_data_frame,
  moi
)
}
\arguments{
\item{response_matrix}{(required) a matrix of raw expression counts. The responses (e.g., genes or proteins) should be in the rows, and the cells should be in the columns. The row names should be the unique IDs of the responses. The matrix can be a standard (dense) matrix or a sparse matrix of class \code{dgCMatrix}, \code{dgRMatrix}, or \code{dgTMatrix}.}

\item{grna_matrix}{(required) a matrix of gRNA expression counts. The gRNAs should be in the rows, and the cells should be in the columns. The row names should be the unique IDs of the gRNAs. The matrix can be a standard (dense) matrix or a sparse matrix of class \code{dgCMatrix}, \code{dgRMatrix}, or \code{dgTMatrix}. (See "Notes" for details about passing a gRNA matrix containing user-specified gRNA-to-cell assignments.)}

\item{covariate_data_frame}{(required) a data frame containing the cell-specific covariates (e.g., sequencing batch, total UMI count, etc.), with cells in the rows and covariates in the columns.}

\item{grna_group_data_frame}{(required) a data frame specifying the "group" to which each individual gRNA belongs. This data frame should contain columns \code{grna_id} and \code{grna_group}, with \code{grna_group} specifying the group membership of a given gRNA in the \code{grna_id} column. Typically, gRNAs that target the same site are grouped into the same gRNA group. Non-targeting gRNAs (if present) should be assigned a gRNA group label of "non-targeting".}

\item{moi}{the MOI of the dataset, either "low" or "high".}
}
\value{
an object of class \code{sceptre_object}
}
\description{
Create a \code{sceptre} object
}
\examples{
#################
# Low MOI example
#################
data(response_matrix_lowmoi) # response-by-cell expression matrix
data(grna_matrix_lowmoi) # gRNA-by-cell expression matrix
data(covariate_data_frame_lowmoi) # cell-by-covariate data frame
data(grna_group_data_frame_lowmoi) # gRNA group information

# 1. create the sceptre object
sceptre_object <- create_sceptre_object(
response_matrix = response_matrix_lowmoi,
grna_matrix = grna_matrix_lowmoi,
covariate_data_frame = covariate_data_frame_lowmoi,
grna_group_data_frame = grna_group_data_frame_lowmoi,
moi = "low")

# 2. plan analysis; this includes setting the formula and obtaining the pairs to analyze
formula_object <- formula(~log(response_n_umis) + log(response_n_nonzero) + bio_rep + p_mito)
discovery_pairs <- generate_all_pairs(response_matrix_lowmoi,
grna_group_data_frame_lowmoi)
sceptre_object <- prepare_analysis(
sceptre_object = sceptre_object,
formula_object = formula_object,
discovery_pairs = discovery_pairs)

# 3. run calibration check
sceptre_object <- run_calibration_check(sceptre_object)
plot_calibration_result(sceptre_object)

##################
# High MOI example
##################
data(response_matrix_highmoi_experimental)
data(grna_matrix_highmoi_experimental)
data(covariate_data_frame_highmoi_experimental)
data(grna_group_data_frame_highmoi_experimental)

# 1. create the sceptre object
sceptre_object <- create_sceptre_object(
response_matrix = response_matrix_highmoi_experimental,
grna_matrix = grna_matrix_highmoi_experimental,
covariate_data_frame = covariate_data_frame_highmoi_experimental,
grna_group_data_frame = grna_group_data_frame_highmoi_experimental,
moi = "high")

# 2. plan analysis; this includes setting the formula and obtaining the pairs to analyze
formula_object <- formula(~log(grna_n_umis) + log(grna_n_nonzero) + log(gene_n_umis) + log(gene_n_nonzero) + batch + p_mito)
data(discovery_pairs_highmoi_experimental)
data(pc_pairs_highmoi_experimental)

sceptre_object <- prepare_analysis(
sceptre_object = sceptre_object,
formula_object = formula_object,
discovery_pairs = discovery_pairs_highmoi_experimental,
positive_control_pairs = pc_pairs_highmoi_experimental,
resampling_mechanism = "permutations",
side = "left")

# 3. run the calibration check
sceptre_object <- run_calibration_check(sceptre_object)
plot_calibration_result(sceptre_object)
}
