% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_data_functions.R
\name{extract_column_from_csc}
\alias{extract_column_from_csc}
\title{Extract column from CSC (compreseed sparse column) representation of matrix}
\usage{
extract_column_from_csc(
  column_no,
  dat,
  indices,
  ind_ptr,
  n_rows,
  zero_based_idx = TRUE,
  row_idxs = NULL
)
}
\arguments{
\item{column_no}{integer vector; the columns to extract from the matrix}

\item{dat}{the vector of UMI counts}

\item{indices}{the "external" indices corresponding to row ids}

\item{ind_ptr}{the "internal" indices used to locate the external indices and dat for a given column}

\item{n_rows}{number of rows of matrix}

\item{zero_based_idx}{(optional boolean) do the indices and ind_ptr vectors use zero-based indexing?}

\item{row_idxs}{(optional) integer vector specifying the rows to extract}
}
\value{
the (dense) columns of the matrix corresponding to columns column_no
}
\description{
Extract column from CSC (compreseed sparse column) representation of matrix
}
