% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_method_functions.R
\name{fit_skew_t}
\alias{fit_skew_t}
\title{Fit skew-t}
\usage{
fit_skew_t(t_nulls, t_star, side = "left")
}
\arguments{
\item{t_nulls}{vector of random sampled statistics from null distribution}

\item{t_star}{numeric, test statistics}

\item{side}{an argument to set test for left-sided, right-sided or both-sided. Default as 'left' and can take 'left', 'right' or 'both'}
}
\value{
A list containing (i) skew_t_fit_success (boolean), (ii) out_p (the p-value), and (iii) skew-t mle (a vector containing the fitted MLEs, NA if fit failed).
}
\description{
Fits a skew-t distribution on a set of resampled test statistics.
}
