% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_gRNA_pairs}
\alias{gene_gRNA_pairs}
\title{Gene-gRNA pairs}
\format{
An object of class \code{data.frame} with 23 rows and 2 columns.
}
\usage{
data(gene_gRNA_pairs)
}
\description{
The gene-gRNA pairs that we seek to test for association. Columns include \code{gene_id} and \code{gRNA_id}.
}
\section{Overview}{

This object is part of the example data in the \code{sceptre} package. The data are taken from the paper "Global Analysis of Enhancer Targets Reveals Convergent Enhancer-Driven Regulatory Modules" by Xie et al., 2019. The authors used CRISPRi-based perturb-seq to target >500 putative enhancers in a population of K562 cells. The example data consist of a subset of the sequenced gRNAs and genes measured on 106,660 cells.
}

\keyword{datasets}
