% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_position_data_frame_grch38}
\alias{gene_position_data_frame_grch38}
\title{Gene position data frame}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 36572 rows and 3 columns.
}
\usage{
data(gene_position_data_frame_grch38)
}
\description{
\code{gene_position_data_frame_grch38} maps each gene to the chromosome on which it is located and to the position of its transcription start site on that chromosome. The data frame was constructed from the GRCh38 reference genome that has shipped with CellRanger since 2020.
}
\keyword{datasets}
