% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_position_data_frame_grch38}
\alias{gene_position_data_frame_grch38}
\alias{gene_position_data_frame_grch37}
\title{Gene position data frames}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 36572 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 57773 rows and 4 columns.
}
\usage{
data(gene_position_data_frame_grch38)

data(gene_position_data_frame_grch37)
}
\description{
\code{gene_position_data_frame_grch38} and \code{gene_position_data_frame_grch37} contain the coordinate and transcription start site position of each gene relative to reference genome GRCh38 and GRCh37, respectively. Both \code{gene_position_data_frame_grch38} and \code{gene_position_data_frame_grch37} were constructed from reference genomes available on the 10x Genomics website. The GRCh38 reference genome has been used by 10x Cell Ranger since 2020.
}
\examples{
head(gene_position_data_frame_grch38)
head(gene_position_data_frame_grch37)
}
\keyword{datasets}
