% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{get_grna_assignments}
\alias{get_grna_assignments}
\title{Get gRNA assignments}
\usage{
get_grna_assignments(sceptre_object, apply_cellwise_qc = FALSE)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object} that has had \code{assign_grnas()} called on it}

\item{apply_cellwise_qc}{a logical value (i.e., \code{TRUE} or \code{FALSE}) indicating whether to return the gRNA-to-cell assignment matrix after cellwise QC has been applied (default \code{FALSE})}
}
\value{
a sparse logical matrix containing the gRNA-to-cell assignments
}
\description{
Obtains the gRNA-to-cell assignments from a \code{sceptre_object}. The output is a sparse logical matrix, with gRNAs in the rows and cells in the columns. A given entry of the matrix is TRUE if the given gRNA is assigned to the given cell.
}
\details{
\itemize{
\item When using the "maximum" assignment strategy, exactly one gRNA is assigned to a given cell. In other words, each column of the gRNA-to-cell assignment matrix contains exactly one TRUE entry.
}
}
