% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{get_grna_assignments}
\alias{get_grna_assignments}
\title{Get gRNA assignments}
\usage{
get_grna_assignments(sceptre_object)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object} that has had \code{assign_grnas()} called on it}
}
\value{
a list containing the gRNA-to-cell assignments
}
\description{
Obtains the gRNA-to-cell assignments from a \code{sceptre_object}. The output is a named list, with each entry corresponding to a different gRNA. The entry for a given gRNA is an integer vector indicating the cells to which that gRNA has been assigned.
}
\details{
The assignments correspond to the original (i.e., non-QC'ed) gRNA expression matrix.
}
