% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{get_response_matrix}
\alias{get_response_matrix}
\alias{get_grna_matrix}
\alias{get_cell_covariates}
\title{Getter functions}
\usage{
get_response_matrix(sceptre_object)

get_grna_matrix(sceptre_object)

get_cell_covariates(sceptre_object)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}
}
\value{
\code{get_response_matrix()} returns the response matrix contained within a \code{sceptre_object}.

\code{get_grna_matrix()} returns the gRNA matrix contained within a \code{sceptre_object}.

\code{get_cell_covariates()} returns the cell covariate data frame contained within a \code{sceptre_object}.
}
\description{
The getter functions (i.e., \verb{get_*}) return of a specified field from a \code{sceptre_object}.
}
\examples{
# 1. create a sceptre_object
directories <- paste0(system.file("extdata", package = "sceptre"),
"/highmoi_example/gem_group_", 1:2)
data(grna_target_data_frame_highmoi)
sceptre_object <- import_data_from_cellranger(
  directories = directories,
  moi = "high",
  grna_target_data_frame = grna_target_data_frame_highmoi
)

# 2. extract data fields from the sceptre_object
response_matrix <- get_response_matrix(sceptre_object)
grna_matrix <- get_grna_matrix(sceptre_object)
cell_covariates <- get_cell_covariates(sceptre_object)
}
