% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_2.R
\name{get_result}
\alias{get_result}
\title{Get result}
\usage{
get_result(sceptre_object, analysis)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}

\item{analysis}{a string indicating the name of the analysis whose results we are querying, one of \code{"run_calibration_check"}, \code{"run_power_check"}, or \code{"run_discovery_analysis"}.}
}
\value{
a data frame containing the results of the analysis
}
\description{
\code{get_result()} returns a data frame containing the result of a calibration check, power analysis, or discovery analysis. See \href{https://timothy-barry.github.io/sceptre-book/sceptre.html#sec-sceptre_write_outputs_to_directory}{Section 8 of the introductory chapter in the manual} for more information about this function.
}
\examples{
library(sceptredata)
data(highmoi_example_data)
data(grna_target_data_frame_highmoi)
# import data
sceptre_object <- import_data(
  response_matrix = highmoi_example_data$response_matrix,
  grna_matrix = highmoi_example_data$grna_matrix,
  grna_target_data_frame = grna_target_data_frame_highmoi,
  moi = "high",
  extra_covariates = highmoi_example_data$extra_covariates,
  response_names = highmoi_example_data$gene_names
)
positive_control_pairs <- construct_positive_control_pairs(sceptre_object)
pc_result <- sceptre_object |>
  set_analysis_parameters(
    side = "left",
    resampling_mechanism = "permutations",
    positive_control_pairs = positive_control_pairs
  ) |>
  assign_grnas(method = "thresholding") |>
  run_qc() |>
  run_power_check() |>
  get_result("run_power_check")
}
