% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{highmoi_example_data}
\alias{highmoi_example_data}
\title{Example high-MOI data}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(highmoi_example_data)
}
\description{
The example high-MOI CRISPRi data are a small simulated dataset modeled on
that of "A genome-wide framework for mapping gene regulation via cellular
genetic screens" by Gasperini et al., 2019. These data contain perturbations
both of gene transcription start sites (TSSs) and enhancers, as well as non-
targeting perturbations.
}
\details{
The example data are stored in a list containing the following components:
\itemize{
\item \code{response_matrix}: the gene-by-cell expression matrix
\item \code{grna_matrix}: the gRNA-by-cell expression matrix
\item \code{extra_covariates}: a data frame containing a single column, \code{batch}, specifying the batch in which each cell was sequenced (\code{batch_1} or \code{batch_2})
\item \code{gene_names}: the human-readable name of each gene
}
}
\keyword{datasets}
