% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_1.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(
  response_matrix,
  grna_matrix,
  grna_target_data_frame,
  moi,
  extra_covariates = data.frame(),
  response_names = NA_character_
)
}
\arguments{
\item{response_matrix}{a matrix of response UMI counts, with responses in rows and cells in columns}

\item{grna_matrix}{a matrix of gRNA UMI counts, with gRNAs in rows and cells in columns}

\item{grna_target_data_frame}{a data frame containing columns \code{grna_id} and \code{grna_target} mapping each individual gRNA to its target}

\item{moi}{a string indicating the MOI of the dataset, either "low" or "high"}

\item{extra_covariates}{(optional) a data frame containing extra covariates (e.g., batch, biological replicate) beyond those that \code{sceptre} can compute}

\item{response_names}{(optional) a vector of human-readable response names; names with the prefix "MT-" are assumed to be mitochondrial genes and are used to compute the covariate \code{response_p_mito}.}
}
\value{
an initialized \code{sceptre_object}
}
\description{
\code{import_data()} imports data from a collection of R objects to create a \code{sceptre_object}. See \href{https://timothy-barry.github.io/sceptre-book/import-data.html#import-data-from-a-collection-of-r-objects}{Chapter 1 of the manual} for detailed information about this function.
}
\examples{
# see example via ?sceptre
}
