% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_1.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(
  response_matrix,
  grna_matrix,
  grna_target_data_frame,
  moi,
  extra_covariates = NULL,
  response_names = NA_character_
)
}
\arguments{
\item{response_matrix}{TBD}

\item{grna_matrix}{TBD}

\item{grna_target_data_frame}{TBD}

\item{moi}{TBD}

\item{extra_covariates}{TBD}

\item{response_names}{TBD}
}
\description{
Import data
}
\examples{
\dontrun{
#################
# Low MOI example
#################
data("lowmoi_example_data")

# 1. create the sceptre object
sceptre_object <- import_data(
response_matrix = lowmoi_example_data$response_matrix,
grna_matrix = lowmoi_example_data$grna_matrix,
extra_covariates = lowmoi_example_data$extra_covariates,
grna_target_data_frame = lowmoi_example_data$grna_target_data_frame,
moi = "low")
print(sceptre_object)

# 2. obtain the discovery and positive control pairs
positive_control_pairs <- construct_positive_control_pairs(sceptre_object)
discovery_pairs <- construct_trans_pairs(sceptre_object = sceptre_object,
positive_control_pairs = positive_control_pairs)

# 3. set the analysis parameters
sceptre_object <- set_analysis_parameters(
sceptre_object = sceptre_object,
discovery_pairs = discovery_pairs,
positive_control_pairs = positive_control_pairs)
print(sceptre_object)

# 4. optional: explicitly assign grnas, run QC
plot_grna_count_distributions(sceptre_object)
sceptre_object <- sceptre_object |> assign_grnas()
plot(sceptre_object)
print(sceptre_object)

sceptre_object <- sceptre_object |> run_qc(p_mito_threshold = 0.1)
plot(sceptre_object)
print(sceptre_object)

# 5. run the calibration check
sceptre_object <- run_calibration_check(sceptre_object)
plot(sceptre_object)
print(sceptre_object)

# 6. run power check
sceptre_object <- run_power_check(sceptre_object)
plot(sceptre_object)
print(sceptre_object)

# 7. run discovery analysis
sceptre_object <- run_discovery_analysis(sceptre_object)
plot(sceptre_object)
print(sceptre_object)

# 8. obtain the results for downstream analysis
write_outputs_to_directory(sceptre_object = sceptre_object, "~/sceptre_outputs/")

##################
# High MOI example
##################
# 1. create the sceptre object from CellRanger output
directories <- paste0(system.file("extdata", package = "sceptre"),
"/highmoi_example/gem_group_", 1:2)
data(grna_target_data_frame_highmoi)
sceptre_object <- import_data_from_cellranger(directories = directories,
moi = "high",
grna_target_data_frame = grna_target_data_frame_highmoi)
print(sceptre_object)

# 2. obtain the response-gRNA group pairs to analyze
positive_control_pairs <- construct_positive_control_pairs(sceptre_object)
discovery_pairs <- construct_cis_pairs(sceptre_object,
positive_control_pairs = positive_control_pairs,
distance_threshold = 5e6)

# 3. set the analysis parameters
sceptre_object <- set_analysis_parameters(
sceptre_object = sceptre_object,
discovery_pairs = discovery_pairs,
positive_control_pairs = positive_control_pairs,
side = "left")
print(sceptre_object)

# 4 (optional) manually assign grnas, run QC
plot_grna_count_distributions(sceptre_object)
sceptre_object <- sceptre_object |> assign_grnas()
plot(sceptre_object)
print(sceptre_object)

sceptre_object <- sceptre_object |> run_qc(p_mito_threshold = 0.1)
plot(sceptre_object)
print(sceptre_object)

# 5. run the calibration check
sceptre_object <- run_calibration_check(sceptre_object)
plot(sceptre_object)
print(sceptre_object)

# 6. (optional) run the power check
sceptre_object <- run_power_check(sceptre_object)
plot(sceptre_object)
print(sceptre_object)

# 7. run discovery analysis
sceptre_object <- run_discovery_analysis(sceptre_object)
plot(sceptre_object)
print(sceptre_object)

# 8. obtain results; write outputs to directory
write_outputs_to_directory(sceptre_object = sceptre_object, "~/sceptre_outputs/")
}
}
