% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functs.R
\name{import_data_from_cellranger}
\alias{import_data_from_cellranger}
\title{Import data from cellranger}
\usage{
import_data_from_cellranger(
  directories,
  moi,
  grna_target_data_frame,
  extra_covariates = NULL
)
}
\arguments{
\item{directories}{TBD}

\item{moi}{TBD}

\item{grna_target_data_frame}{TBD}
}
\value{
a sceptre_object initialized
}
\description{
Import data from cellranger
}
\examples{
\dontrun{
# 1. create the sceptre object from CellRanger output
directories <- paste0(system.file("extdata", package = "sceptre"),
                     "/highmoi_example/gem_group_", 1:2)
data(grna_target_data_frame_highmoi)

# 2. simulate an additional covariate
cell_type <- sample(x = paste0("type_", 1:3),
size = 45919,
replace = TRUE) |> factor()
extra_covariates <- data.frame(cell_type = cell_type)

# 3. initialize the sceptre_object
sceptre_object <- import_data_from_cellranger(directories = directories,
                                              moi = "high",
                                              grna_target_data_frame = grna_target_data_frame_highmoi,
                                              extra_covariates = extra_covariates)
}
}
