% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functs.R
\name{import_data_from_cellranger}
\alias{import_data_from_cellranger}
\title{Import data from Cell Ranger}
\usage{
import_data_from_cellranger(
  directories,
  moi,
  grna_target_data_frame,
  extra_covariates = data.frame(),
  use_ondisc = FALSE,
  directory_to_write = NULL
)
}
\arguments{
\item{directories}{a character vector of file paths to directories containing the output of one or more calls to Cell Ranger count. Each directory should contain the files \code{matrix.mtx.gz} and \code{features.tsv.gz} (and optionally \code{barcodes.tsv.gz}).}

\item{moi}{a string indicating the MOI of the dataset, either "low" or "high".}

\item{grna_target_data_frame}{a data frame containing columns \code{grna_id} and \code{grna_target} mapping each individual gRNA to its target. Non-targeting gRNAs should be assigned a label of "non-targeting". Optionally, \code{grna_target_data_frame} can contain columns \code{chr}, \code{start}, and \code{end}, giving the chromosome, start coordinate, and end coordiante, respectively, of each gRNA. Additionally, \code{grna_target_data_frame} can contain the column \code{vector_id} specifying the vector to which a given gRNA belongs.}

\item{extra_covariates}{(optional) a data frame containing extra covariates (e.g., batch, biological replicate) beyond those that \code{sceptre} can compute.}

\item{use_ondisc}{(optional; default \code{FALSE}) a logical indicating whether to store the expression data in a disk-backed \code{ondisc} matrix (\code{TRUE}) or an in-memory sparse matrix (\code{FALSE}).}

\item{directory_to_write}{(optional) a string indicating the directory in which to write the backing \code{.odm} files (must be specified if \code{use_ondisc} is set to \code{TRUE}).}
}
\value{
an initialized \code{sceptre_object}
}
\description{
\code{import_data_from_cellranger()} imports data from the output of one or more calls to Cell Ranger count. Each directory supplied as an input to this function should be in feature-barcode format, containing the files \code{features.tsv.gz} and  \code{matrix.mtx.gz} (and optionally \code{barcodes.tsv.gz}). Users can create either a standard \code{sceptre} object or an \code{ondisc}-backed \code{sceptre} object; the latter is more appropriate for large-scale data. See \href{https://timothy-barry.github.io/sceptre-book/sceptre.html#sec-whole_game_import_data}{the introductory chapter} or \href{https://timothy-barry.github.io/sceptre-book/import-data.html}{Chapter 1} of the manual for more information about this function.
}
\examples{
data(grna_target_data_frame_highmoi)
directories <- paste0(
  system.file("extdata", package = "sceptre"),
  "/highmoi_example/gem_group_", c(1, 2)
)

# 1. create a standard sceptre_object from Cell Ranger output
sceptre_object <- import_data_from_cellranger(
  directories = directories,
  moi = "high",
  grna_target_data_frame = grna_target_data_frame_highmoi,
)

# 2. create an ondisc-backed sceptre_object from Cell Ranger output
if (requireNamespace("ondisc", quietly = TRUE) &&
    utils::packageVersion("ondisc") >= package_version("1.2.0")) {
  sceptre_object <- import_data_from_cellranger(
    directories = directories,
    moi = "high",
    grna_target_data_frame = grna_target_data_frame_highmoi,
    use_ondisc = TRUE,
    directory_to_write = tempdir()
  )
}
}
