% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functions.R
\name{initialize_directories}
\alias{initialize_directories}
\title{Initialze directories}
\usage{
initialize_directories(storage_location)
}
\arguments{
\item{storage_location}{file path to a directory in which to store the intermediate and final results}
}
\value{
named vector containing the file paths of the gene_precomp_dir, gRNA_precomp_dir, results_dir, and log_dir.
}
\description{
Initialze directories
}
