% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lowmoi_example_data}
\alias{lowmoi_example_data}
\title{Example low-MOI data}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(lowmoi_example_data)
}
\description{
The example low MOI data are taken from the paper "Characterizing the molecular regulation of inhibitory immune checkpoints with multimodal single-cell screens" by Papalexi et al., 2021. The authors used CRISPRko to target 26 genes in 20,729 THP-1 cells and measured the effects of these perturbations via single-cell sequencing. The authors designed three or four gRNAs to target each gene as well as a library of nine non-targeting gRNAs. The genes were downsampled for inclusion in the example data.
}
\details{
The example data are stored in a list containing four components:
\itemize{
\item \code{response_matrix}: a gene-by-cell expression matrix
\item \code{grna_matrix}: a gRNA-by-cell expression matrix
\item \code{grna_target_data_frame}:  a data frame containing the columns \code{grna_id} (ID of an individual gRNA) and \code{grna_target} (genomic target of the gRNA)
\item \code{extra_covariates}: a data frame with a single column, \code{bio_rep}, specifying the biological replicate of each cell (\code{rep_1}, \code{rep_2}, or \code{rep_3})
}
}
\keyword{datasets}
