% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_qq_plot}
\alias{make_qq_plot}
\title{Make qq-plot}
\usage{
make_qq_plot(
  p_values,
  ci_level = 0.95,
  subsampling_factor = 1,
  point_col = "royalblue4"
)
}
\arguments{
\item{p_values}{Fill in}

\item{ci_level}{Fill in}

\item{subsampling_factor}{Fill in}

\item{point_col}{Fill in}
}
\value{

}
\description{
Make qq-plot
}
