% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{make_volcano_plot}
\alias{make_volcano_plot}
\title{Make volcano plot}
\usage{
make_volcano_plot(
  discovery_result,
  alpha = 0.1,
  x_limits = c(-1.5, 1.5),
  transparency = 0.5,
  point_size = 0.9,
  multiple_testing_correction = "BH"
)
}
\arguments{
\item{discovery_result}{(required) the output of a call to \code{run_sceptre_lowmoi} with \code{calibration_check} set to \code{FALSE}}

\item{alpha}{(optional; default 0.1) the nominal type-I error control level}

\item{x_limits}{a vector of length two giving the lower and upper limits of the volcano plot}

\item{transparency}{a number in the range (0, 1) indicating the transparency of the points}

\item{point_size}{a positive number indicating the size of the points}

\item{multiple_testing_correction}{(optional; default "BH") the multiple testing correction method (one of "BH", "bonferroni", "BY", or "holm")}
}
\value{
a ggplot2 object
}
\description{
The \code{make_volcano_plot()} function creates a volcano plot of the discovery results. Each point in the plot corresponds to a response-gRNA group pair; the estimated log-2 fold change of the pair is plotted on the x-axis, and the (negative log-10 transformed) p-value is plotted on the y-axis.
}
\details{
The horizontal dashed line indicates the multiple testing threshold. Points above the dashed line (which are colored in purple) are called as discoveries; points below the dashed line (which are colored in blue), meanwhile, are called as insignificant.
}
\examples{
# See the example in the run_sceptre_lowmoi help file.
?run_sceptre_lowmoi
}
