% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{obtain_discovery_set}
\alias{obtain_discovery_set}
\title{Obtain discovery set}
\usage{
obtain_discovery_set(
  discovery_result,
  alpha = 0.1,
  multiple_testing_correction = "BH"
)
}
\value{
a subset of \code{discovery_result} containing the discoveries.
}
\description{
The \code{obtain_discovery_set()} function returns the discovery set from an input discovery result.
}
\examples{
# See the example in the run_sceptre_lowmoi help file.
?run_sceptre_lowmoi
}
