% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_assign_grnas}
\alias{plot_assign_grnas}
\title{Plot the number of gRNAs per cell}
\usage{
plot_assign_grnas(
  sceptre_object,
  n_grnas_to_plot = 2L,
  grnas_to_plot = NULL,
  transparency = 0.8,
  point_size = 0.9,
  return_indiv_plots = FALSE,
  n_max_0_grna_unprtb_plot = 1000
)
}
\arguments{
\item{n_grnas_to_plot}{(optional; default \code{2}) the number of different gRNAs shown in the plots of gRNA count versus cell assigment.}

\item{grnas_to_plot}{(optional; default \code{NULL}) the names of specific gRNAs to plot; if \code{NULL} then gRNAs are chosen at random.}

\item{return_indiv_plots}{(optional; default \code{FALSE}) return a single combined plot (\code{TRUE}) or the individual plots in a list (\code{FALSE})?}

\item{n_max_0_grna_unprtb_plot}{(optional; default \code{1000}) there may be many cells with a gRNA count of 0 in the unperturbed group. This can slow down the plotting without adding useful information, so at most \code{n_max_0_grna_unprtb_plot} points from this group are plotted. Setting this to \code{Inf} will guarantee no downsampling occurs.}
}
\value{
a single \code{cowplot} object containing the combined panels (if \code{return_indiv_plots} is set to \code{TRUE}) or a list of the individual panels (if \code{return_indiv_plots} is set to \code{FALSE}).
}
\description{
Plot the number of gRNAs per cell
}
