% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_assign_grnas}
\alias{plot_assign_grnas}
\title{Plot the outcome of the gRNA-to-cell assignment step}
\usage{
plot_assign_grnas(
  sceptre_object,
  n_grnas_to_plot = 3L,
  grnas_to_plot = NULL,
  point_size = 0.9,
  transparency = 0.8,
  n_max_0_grna_unprtb_plot = 1000,
  return_indiv_plots = FALSE
)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object} that has had \code{assign_gras} called on it.}

\item{n_grnas_to_plot}{(optional; default \code{3}) the number of different gRNAs shown in the plots of gRNA count versus cell assigment.}

\item{grnas_to_plot}{(optional; default \code{NULL}) the names of specific gRNAs to plot; if \code{NULL} then the gRNAs are chosen at random.}

\item{point_size}{(optional; default \code{0.9}) the size of the individual points in the plot.}

\item{transparency}{(optional; default \code{0.8}) the transparency of the individual points in the plot.}

\item{n_max_0_grna_unprtb_plot}{(optional; default \code{1000}) there may be many cells with a gRNA count of 0 in the unperturbed group. This can slow down the plotting without adding useful information, so at most \code{n_max_0_grna_unprtb_plot} points from this group are plotted. Setting this to \code{Inf} will guarantee no downsampling occurs.}

\item{return_indiv_plots}{(optional; default \code{FALSE}) if \code{FALSE} then a list of \code{ggplot} is returned; if \code{TRUE} then a single \code{cowplot} object is returned.}
}
\value{
a single \code{cowplot} object containing the combined panels (if \code{return_indiv_plots} is set to \code{TRUE}) or a list of the individual panels (if \code{return_indiv_plots} is set to \code{FALSE}).
}
\description{
Plot the outcome of the gRNA-to-cell assignment step
}
\examples{
# A full example can be found at \code{?sceptre}
}
