% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_covariates}
\alias{plot_covariates}
\title{Plot covariates}
\usage{
plot_covariates(
  sceptre_object,
  response_n_umis_range = c(0.01, 0.99),
  response_n_nonzero_range = c(0.01, 0.99),
  p_mito_threshold = 0.2,
  return_indiv_plots = FALSE
)
}
\arguments{
\item{sceptre_object}{any initialized \code{sceptre_object}. \code{plot_covariates()} can be called at any point in the pipeline after \code{import_data()}.}

\item{response_n_umis_range}{(optional; default \code{c(0.01, 0.99)}) a length-2 vector of quantiles indicating the location at which to draw vertical lines on the \code{response_n_umis} histogram}

\item{response_n_nonzero_range}{(optional; default \code{c(0.01, 0.99)}) a length-2 vector of quantiles indicating the location at which to draw vertical lines on the \code{response_n_nonzero} histogram}

\item{p_mito_threshold}{(optional; default \code{0.2}) a single numeric value in the interval [0,1] specifying the location at which to draw a vertical line on the \code{response_p_mito} histogram. Note that \code{p_mito_threshold} is an absolute number rather than a percentile.}

\item{return_indiv_plots}{(optional; default \code{FALSE}) if \code{FALSE}, then a list of \code{ggplot} objects is returned; if \code{TRUE} then a single \code{cowplot} object is returned.}
}
\value{
a single \code{cowplot} object containing the combined panels (if \code{return_indiv_plots} is set to \code{TRUE}) or a list of the individual panels (if \code{return_indiv_plots} is set to \code{FALSE})
}
\description{
\code{plot_covariates()} creates histograms of several cell-specific covariates: \code{response_n_nonzero}, \code{response_n_umis}, and (if applicable) \code{response_p_mito}.
}
\details{
To help guide the selection of QC thresholds, \code{plot_covariates()} plots candidate QC thresholds as vertical lines on the histograms. If \code{run_qc()} has been called on the \code{sceptre_object}, then the QC thresholds plotted on the histograms are those contained within the \code{sceptre_object}.
}
\examples{
library(sceptredata)
data(highmoi_example_data)
data(grna_target_data_frame_highmoi)
import_data(
  response_matrix = highmoi_example_data$response_matrix,
  grna_matrix = highmoi_example_data$grna_matrix,
  grna_target_data_frame = grna_target_data_frame_highmoi,
  moi = "high",
  extra_covariates = highmoi_example_data$extra_covariates,
  response_names = highmoi_example_data$gene_names
) |> plot_covariates(p_mito_threshold = 0.07)
}
