% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_covariates}
\alias{plot_covariates}
\title{Plot covariates}
\usage{
plot_covariates(
  sceptre_object,
  response_n_umis_range = c(0.01, 0.99),
  response_n_nonzero_range = c(0.01, 0.99),
  p_mito_threshold = 0.2,
  return_indiv_plots = FALSE
)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}

\item{response_n_umis_range}{(optional; default \code{c(0.01, 0.99)}) a length-2 vector of quantiles indicating the location at which to draw vertical lines on the \code{response_n_umis} histogram}

\item{response_n_nonzero_range}{(optional; default \code{c(0.01, 0.99)}) a length-2 vector of quantiles indicating the location at which to draw vertical lines on the \code{response_n_nonzero} histogram}

\item{p_mito_threshold}{(optional; default \code{0.2}) a single numeric value in the interval [0,1] specifying the location at which to draw a vertical line on the \code{response_p_mito} histogram. Note that \code{p_mito_threshold} is an absolute number rather than a percentile.}

\item{return_indiv_plots}{(optional; default \code{FALSE}) if \code{FALSE}, then a list of \code{ggplot} objects is returned; if \code{TRUE} then a single \code{cowplot} object is returned.}
}
\value{
a single \code{cowplot} object containing the combined panels (if \code{return_indiv_plots} is set to \code{TRUE}) or a list of the individual panels (if \code{return_indiv_plots} is set to \code{FALSE})
}
\description{
\code{plot_covariates()} creates a histogram of the covariates \code{response_n_nonzero}, \code{response_n_umis}, and (if applicable) \code{response_p_mito}. Cellwise QC removes cells that lie in the extreme right tail of the \code{response_p_mito} distribution or that lie in the extreme left \emph{or} right tail of the \code{response_n_nonzero} or \code{response_n_umis} distribution. To help guide the selection of QC thresholds, \code{plot_covariates()} plots candidate QC thresholds as vertical lines on the histograms. The optional arguments \code{response_n_nonzero_range}, \code{response_n_umis_range}, and \code{p_mito_threshold} control the location of these candidate QC thresholds. \code{response_n_nonzero_range} (resp., \code{response_n_umis_range}) is a length-two vector of quantiles (default: \code{c(0.01, 0.99)}) indicating the location at which to draw candidate QC thresholds on the \code{response_n_nonzero} (resp., \code{response_n_umis}) histogram. Next, \code{p_mito_threshold} is a single numeric value in the interval [0,1] specifying the location at which to draw a candidate QC threshold on the \code{response_p_mito} plot.
}
\note{
If \code{run_qc()} has already been called on the \code{sceptre_object}, then the parameters \code{response_n_umis_range}, \code{response_n_nonzero_range}, and \code{p_mito_threshold} are set to the corresponding parameters within the \code{sceptre_object}.
}
\examples{
library(sceptredata)
data(highmoi_example_data)
data(grna_target_data_frame_highmoi)
import_data(
  response_matrix = highmoi_example_data$response_matrix,
  grna_matrix = highmoi_example_data$grna_matrix,
  grna_target_data_frame = grna_target_data_frame_highmoi,
  moi = "high",
  extra_covariates = highmoi_example_data$extra_covariates,
  response_names = highmoi_example_data$gene_names
) |> plot_covariates(p_mito_threshold = 0.07)
}
