% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_run_discovery_analysis}
\alias{plot_run_discovery_analysis}
\title{Plot run discovery analysis}
\usage{
plot_run_discovery_analysis(
  sceptre_object,
  x_limits = c(-1.5, 1.5),
  point_size = 0.55,
  transparency = 0.8,
  return_indiv_plots = FALSE
)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object} that has had \code{run_discovery_analysis} called on it}

\item{x_limits}{(optional; default \code{c(-1.5, 1.5)}) a numeric vector of length 2 giving the lower and upper limits of the x-axis (corresponding to log-2 fold change) for the "Discovery volcano plot" panel}

\item{point_size}{(optional; default \code{0.55}) the size of the individual points in the plot}

\item{transparency}{(optional; default \code{0.8}) the transparency of the individual points in the plot}

\item{return_indiv_plots}{(optional; default \code{FALSE}) if \code{FALSE} then a list of \code{ggplot} is returned; if \code{TRUE} then a single \code{cowplot} object is returned.}
}
\value{
a single \code{cowplot} object containing the combined panels (if \code{return_indiv_plots} is set to \code{TRUE}) or a list of the individual panels (if \code{return_indiv_plots} is set to \code{FALSE})
}
\description{
\code{plot_run_discovery_analysis()} creates a visualization of the outcome of the discovery analysis. See \href{https://timothy-barry.github.io/sceptre-book/sceptre.html#sec-sceptre_run_discovery_analysis}{Section 7 of the introductory chapter in the manual} for guidance on interpreting this visual.
}
\examples{
library(sceptredata)
data(highmoi_example_data)
data(grna_target_data_frame_highmoi)
# import data
sceptre_object <- import_data(
 response_matrix = highmoi_example_data$response_matrix,
 grna_matrix = highmoi_example_data$grna_matrix,
 grna_target_data_frame = grna_target_data_frame_highmoi,
 moi = "high",
 extra_covariates = highmoi_example_data$extra_covariates,
 response_names = highmoi_example_data$gene_names
)
positive_control_pairs <- construct_positive_control_pairs(sceptre_object)
discovery_pairs <- construct_cis_pairs(sceptre_object,
                                      positive_control_pairs = positive_control_pairs,
                                      distance_threshold = 5e6)
sceptre_object |>
set_analysis_parameters(
   side = "left",
   discovery_pairs = discovery_pairs,
   resampling_mechanism = "permutations",
 ) |>
 assign_grnas(method = "thresholding") |>
 run_qc() |>
 run_calibration_check(
    parallel = TRUE,
    n_processors = 2
 ) |>
 run_discovery_analysis(
    parallel = TRUE,
    n_processors = 2
 ) |>
 plot_run_discovery_analysis()
}
