% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_run_discovery_analysis}
\alias{plot_run_discovery_analysis}
\title{Plot run discovery analysis}
\usage{
plot_run_discovery_analysis(
  sceptre_object,
  return_indiv_plots = FALSE,
  x_limits = c(-1.5, 1.5),
  transparency = 0.8,
  point_size = 0.55
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{return_indiv_plots}{TBD}

\item{x_limits}{TBD}

\item{transparency}{TBD}

\item{point_size}{TBD}
}
\description{
Plot run discovery analysis
}
