% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_run_qc}
\alias{plot_run_qc}
\title{Plot run QC}
\usage{
plot_run_qc(
  sceptre_object,
  downsample_pairs = 10000L,
  point_size = 0.55,
  transparency = 0.8,
  return_indiv_plots = FALSE
)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object} that has had \code{run_qc} called on it}

\item{downsample_pairs}{(optional; default \code{10000}) integer; the discovery pairs in \code{sceptre_object} are downsampled to just \code{downsample_pairs} pairs, speeding up the "PairWise QC" plot in the lower panel}

\item{point_size}{(optional; default \code{0.55}) the size of the individual points in the plot.}

\item{transparency}{(optional; default \code{0.8}) the transparency of the individual points in the plot.}

\item{return_indiv_plots}{(optional; default \code{FALSE}) if \code{FALSE} then a list of \code{ggplot} is returned; if \code{TRUE} then a single \code{cowplot} object is returned.}
}
\description{
\code{plot_run_qc()} visualizes the outcome of the QC step.
}
\examples{
# A full example can be found at ?sceptre;
# `plot_run_qc()` is dispatched when
# `plot()` is called on the `sceptre_object`
# in step 4 (the run qc step).
}
