% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_run_qc}
\alias{plot_run_qc}
\title{Plot run QC}
\usage{
plot_run_qc(
  sceptre_object,
  return_indiv_plots = FALSE,
  transparency = 0.8,
  point_size = 0.55
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{return_indiv_plots}{TBD}

\item{transparency}{TBD}

\item{point_size}{TBD}
}
\description{
Plot run QC
}
