% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_skew_t}
\alias{plot_skew_t}
\title{Plot resampled test statistics over skew-t distribution}
\usage{
plot_skew_t(resampled_zvalues, original_zvalue, dp, interval = c(-4, 4))
}
\arguments{
\item{resampled_zvalues}{the set of resampled z-values}

\item{original_zvalue}{the z value of the original negative binomial fit}

\item{dp}{the skew-t fit MLE}

\item{interval}{(optional; default c(-4,4)) interval over which to plot the skew-t distribution}
}
\value{
a ggplot object containing the plot
}
\description{
Plots the fitted skew-t distribution, along with the resampled (null) test statistics and real test statistic.
}
\examples{
data(expressions)
data(gRNA_indicators)
data(covariate_matrix)
result <- run_sceptre_gRNA_gene_pair(expressions = expressions,
gRNA_indicators = gRNA_indicators,
covariate_matrix = covariate_matrix,
reduced_output = FALSE,
seed = 4,
verbose = FALSE)
plot_skew_t(result$resampled_z_values, result$z_value, result$skew_t_mle)
}
