% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ondisc_functs.R
\name{read_ondisc_backed_sceptre_object}
\alias{read_ondisc_backed_sceptre_object}
\title{Read ondisc-backed sceptre object}
\usage{
read_ondisc_backed_sceptre_object(
  sceptre_object_fp,
  response_odm_file_fp,
  grna_odm_file_fp
)
}
\arguments{
\item{sceptre_object_fp}{file path to a \code{sceptre_object.rds} file}

\item{response_odm_file_fp}{file path to a backing \code{.odm} file for the response modality}

\item{grna_odm_file_fp}{file path to a backing \code{.odm} file for the gRNA modality}
}
\value{
an odm-backed \code{sceptre_object}
}
\description{
Reads and initializes a \code{sceptre_object} from backing .odm files.
}
