% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ondisc_functs.R
\name{read_ondisc_backed_sceptre_object}
\alias{read_ondisc_backed_sceptre_object}
\alias{write_ondisc_backed_sceptre_object}
\title{Read ondisc-backed sceptre object}
\usage{
read_ondisc_backed_sceptre_object(
  sceptre_object_fp,
  response_odm_file_fp,
  grna_odm_file_fp
)

write_ondisc_backed_sceptre_object(sceptre_object, directory_to_write)
}
\arguments{
\item{sceptre_object_fp}{file path to a \code{sceptre_object.rds} file}

\item{response_odm_file_fp}{file path to a backing \code{.odm} file for the response modality}

\item{grna_odm_file_fp}{file path to a backing \code{.odm} file for the gRNA modality}

\item{sceptre_object}{a \code{sceptre_object}}

\item{directory_to_write}{directory in which to write \code{sceptre_object.rds}}
}
\value{
an ondisc-backed \code{sceptre_object}
}
\description{
Reads and initializes a \code{sceptre_object} from backing .odm files.

Write an ondisc-backed sceptre object to disk.
}
