% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_2.R
\name{run_calibration_check}
\alias{run_calibration_check}
\title{Run calibration check}
\usage{
run_calibration_check(
  sceptre_object,
  output_amount = 1,
  n_calibration_pairs = NULL,
  calibration_group_size = NULL,
  print_progress = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}

\item{output_amount}{(optional; default \code{1}) an integer taking values 1, 2, or 3 specifying the amount of information to return}

\item{n_calibration_pairs}{(optional) the number of negative control pairs to construct and test for association}

\item{calibration_group_size}{(optional) the number of negative control gRNAs to put into each negative control target}

\item{print_progress}{(optional; default \code{TRUE}) a logical indicating whether to print progress updates}

\item{parallel}{(optional; default \code{FALSE}) a logical indicating whether to run the function in parallel}
}
\value{
an updated \code{sceptre_object} in which the calibration check has been carried out
}
\description{
\code{run_calibration_check()} runs the calibration check. See \href{https://timothy-barry.github.io/sceptre-book/run-calibration-check.html}{Chapter 5 of the manual} for detailed information about this function.
}
\examples{
# see example via ?sceptre
}
