% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_2.R
\name{run_discovery_analysis}
\alias{run_discovery_analysis}
\title{Run discovery analysis}
\usage{
run_discovery_analysis(
  sceptre_object,
  output_amount = 1,
  print_progress = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{output_amount}{TBD}

\item{print_progress}{TBD}

\item{parallel}{TBD}
}
\description{
Run discovery analysis
}
