% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functions.R
\name{run_gene_precomputation_at_scale_round_1}
\alias{run_gene_precomputation_at_scale_round_1}
\title{Run_gene_precomputation_at_scale_round_1}
\usage{
run_gene_precomputation_at_scale_round_1(
  pod_id,
  gene_precomp_dir,
  cell_gene_expression_matrix,
  ordered_gene_ids,
  covariate_matrix,
  regularization_amount,
  cell_subset,
  log_dir
)
}
\arguments{
\item{pod_id}{pod id}

\item{gene_precomp_dir}{location of the gene precomputation directory}

\item{cell_gene_expression_matrix}{the cell-by-gene expression matrix}

\item{ordered_gene_ids}{the ordered gene ids (i.e., column names of the above matrix)}

\item{covariate_matrix}{the cell-specific covariate matrix}

\item{cell_subset}{(optional) vector indicating subset of cells to use}

\item{log_dir}{(optional) directory in which to sink the log file}
}
\description{
This function runs the first round of gene precomputations. In particular, it computes the raw dispersion estimate and log geometric mean of each gene. It saves the results in the gene_precomp directory.
}
