% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{run_gene_precomputation_at_scale_round_1}
\alias{run_gene_precomputation_at_scale_round_1}
\title{Run_gene_precomputation_at_scale_round_1}
\usage{
run_gene_precomputation_at_scale_round_1(
  pod_id,
  gene_precomp_dir,
  expression_matrix,
  covariate_matrix,
  regularization_amount,
  log_dir
)
}
\arguments{
\item{pod_id}{pod id}

\item{gene_precomp_dir}{location of the gene precomputation directory}

\item{expression_matrix}{an ondisc_matrix representing the expression data}

\item{covariate_matrix}{the cell-specific covariate matrix}

\item{regularization_amount}{amount of regularization to apply to the thetas}

\item{log_dir}{directory in which to sink the log file}
}
\description{
NOTE: DO NOT USE THIS FUNCTION DIRECTLY. INSTEAD, USE THE "AT-SCALE" BASH SCRIPT,
WHICH IS DESCRIBED HERE: https://timothy-barry.github.io/sceptre/articles/sceptre-at-scale.html
}
\details{
This function runs the first round of gene precomputations. In particular, it computes the raw dispersion estimate
and log geometric mean of each gene. It saves the results in the gene_precomp directory.
}
