% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{run_gene_precomputation_at_scale_round_2}
\alias{run_gene_precomputation_at_scale_round_2}
\title{Run gene precomputation at scale round 2}
\usage{
run_gene_precomputation_at_scale_round_2(
  pod_id,
  gene_precomp_dir,
  expression_matrix,
  covariate_matrix,
  regularization_amount,
  log_dir
)
}
\arguments{
\item{pod_id}{pod id}

\item{gene_precomp_dir}{gene precomp dir}

\item{covariate_matrix}{covariate matrix}

\item{log_dir}{directory in which to sink the logs}

\item{cell_gene_expression_matrix}{cell gene expression matrix}

\item{ordered_gene_ids}{ordered gene ids}

\item{cell_subset}{cell subset}
}
\description{
Runs the second round of gene precomputations.
}
