% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{run_gene_precomputation_at_scale_round_2}
\alias{run_gene_precomputation_at_scale_round_2}
\title{Run gene precomputation at scale round 2}
\usage{
run_gene_precomputation_at_scale_round_2(
  pod_id,
  gene_precomp_dir,
  gene_matrix,
  covariate_matrix,
  regularization_amount,
  log_dir
)
}
\arguments{
\item{pod_id}{pod id}

\item{gene_precomp_dir}{gene precomp dir}

\item{gene_matrix}{the gene expression matrix, stored as an ondisc_matrix}

\item{covariate_matrix}{the cell-specific matrix of technical factors, which ideally contains the following covariates: log-transformed gene library size (numeric), log-transformed gRNA library size (numeric), percent mitochondrial reads (numeric), and batch (factor). The rows (i.e., cell barcodes) should be named}

\item{regularization_amount}{non-negative number specifying the amount of regularization to apply to the negative binomial dispersion parameter estimates}

\item{log_dir}{directory in which to sink the logs}
}
\description{
Runs the second round of gene precomputations.
}
