% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functions.R
\name{run_gRNA_gene_pair_analysis_at_scale}
\alias{run_gRNA_gene_pair_analysis_at_scale}
\title{Run gRNA-gene pair analysis at scale}
\usage{
run_gRNA_gene_pair_analysis_at_scale(
  pod_id,
  gene_precomp_dir,
  gRNA_precomp_dir,
  results_dir,
  cell_gene_expression_matrix,
  ordered_gene_ids,
  gRNA_indicator_matrix_fp,
  covariate_matrix,
  cell_subset = NULL,
  B = 500,
  log_dir = NULL,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Run gRNA-gene pair analysis at scale
}
