% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{run_gRNA_gene_pair_analysis_at_scale}
\alias{run_gRNA_gene_pair_analysis_at_scale}
\title{Run gRNA-gene pair analysis at scale}
\usage{
run_gRNA_gene_pair_analysis_at_scale(
  pod_id,
  gene_precomp_dir,
  gRNA_precomp_dir,
  results_dir,
  log_dir,
  gene_matrix,
  gRNA_matrix,
  covariate_matrix,
  regularization_amount,
  side,
  B,
  full_output
)
}
\arguments{
\item{pod_id}{id of the pod to compute}

\item{gene_precomp_dir}{the directory containing the results of the gene precomputation}

\item{gRNA_precomp_dir}{the directory containing the results of the gRNA precomputation}

\item{results_dir}{directory in which to store the results}

\item{log_dir}{(optional) directory in which to sink the log file}

\item{gene_matrix}{the expression matrix, stored as an ondisc_matrix}

\item{gRNA_matrix}{the matrix of perturbations, stored as an ondisc_matrix}

\item{covariate_matrix}{the cell-covariate matrix}

\item{regularization_amount}{amount of regularuzation to apply to thetas}

\item{side}{(optional, default left) sidedness of test}

\item{B}{number of bootstrap resamples (default 500)}

\item{seed}{(optional) seed to pass to the randomization algorithm}
}
\description{
NOTE: DO NOT USE THIS FUNCTION DIRECTLY. INSTEAD, USE THE "AT-SCALE" BASH SCRIPT,
WHICH IS DESCRIBED HERE: https://timothy-barry.github.io/sceptre/articles/sceptre-at-scale.html
}
\details{
Runs the gene-gRRNA pair ananysis across an entire pod of gene-gRNA pairs.
}
