% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_method_functions.R
\name{run_gRNA_precomputation}
\alias{run_gRNA_precomputation}
\title{Run gRNA precomputation}
\usage{
run_gRNA_precomputation(gRNA_indicators, covariate_matrix)
}
\arguments{
\item{gRNA_indicators}{a vector of gRNA indicators}

\item{covariate_matrix}{the cell-specific covariate matrix}
}
\value{
the fitted probabilities
}
\description{
This function runs the precomputation for a given gRNA.
}
