% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functions.R
\name{run_gRNA_precomputation_at_scale}
\alias{run_gRNA_precomputation_at_scale}
\title{Run gRNA precomputation at scale}
\usage{
run_gRNA_precomputation_at_scale(
  gRNA_ids,
  gRNA_indicator_matrix_fp,
  covariate_matrix,
  precomp_matrix_fp,
  cell_subset = NULL
)
}
\arguments{
\item{gRNA_ids}{the ids of gRNAs on which to compute the precomputation}

\item{gRNA_indicator_matrix_fp}{file path the the gRNA indicator matrix (assumed to be stored in .fst format)}

\item{covariate_matrix}{the covariate matrix}

\item{precomp_matrix_fp}{the name of the file (including the file path) in which to save the precomputation result}

\item{cell_subset}{(optional) an integer vector containing the cells to use in the precomputation}
}
\description{
Run gRNA precomputation at scale
}
\examples{
gRNA_dict <- read.fst("/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/precomp/gRNA/gRNA_dictionary.fst")
gRNA_ids <- filter(gRNA_dict, pod_id == 1) \%>\% pull(id) \%>\% head(20)
precomp_matrix_fp <- as.character((filter(gRNA_dict, pod_id == 1) \%>\% pull(file_name))[1])
gRNA_indicator_matrix_fp <- "/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/processed/gRNA_indicators.fst"
covariate_matrix <- read.fst("/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/processed/cell_covariate_model_matrix.fst")
run_gRNA_precomputation_at_scale(gRNA_ids[1:20], gRNA_indicator_matrix_fp, covariate_matrix, precomp_matrix_fp)
}
