% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functions.R
\name{run_gRNA_precomputation_at_scale}
\alias{run_gRNA_precomputation_at_scale}
\title{Run gRNA precomputation at scale}
\usage{
run_gRNA_precomputation_at_scale(
  pod_id,
  gRNA_precomp_dir,
  gRNA_indicator_matrix_fp,
  covariate_matrix,
  cell_subset,
  log_dir
)
}
\arguments{
\item{pod_id}{ID of the pod for which to do the precomputation}

\item{gRNA_precomp_dir}{file path to the gRNA precomputation directory}

\item{gRNA_indicator_matrix_fp}{filepath to the gRNA indicator matrix}

\item{covariate_matrix}{the cell-specific covariate matrix}

\item{cell_subset}{an integer vector specifying the cells on which to fit the model}
}
\description{
This function runs the gRNA precomputation on a selected "pod" of gRNAs (as identified by the pod_id). It stores the result in the gRNA precomputation directory.
}
\examples{
pod_id <- 1
gRNA_precomp_dir <- "/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/precomp/gRNA"
gRNA_indicator_matrix_fp <- "/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/processed/gRNA_indicators.fst"
covariate_matrix <- read.fst("/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/processed/cell_covariate_model_matrix.fst")
cell_subset <- readRDS("/Volumes/tims_new_drive/research/sceptre_files/data/gasperini/processed/cells_to_keep.rds")
}
