% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_up_functs.R
\name{run_gRNA_precomputation_at_scale}
\alias{run_gRNA_precomputation_at_scale}
\title{Run gRNA precomputation at scale}
\usage{
run_gRNA_precomputation_at_scale(
  pod_id,
  gRNA_precomp_dir,
  gRNA_matrix,
  covariate_matrix,
  log_dir,
  B,
  seed
)
}
\arguments{
\item{pod_id}{ID of the pod for which to do the precomputation}

\item{gRNA_precomp_dir}{file path to the gRNA precomputation directory}

\item{gRNA_matrix}{the matrix of perturbations, stored as an ondisc_matrix}

\item{covariate_matrix}{the cell-specific covariate matrix}

\item{log_dir}{file path to the log directory}
}
\description{
NOTE: DO NOT USE THIS FUNCTION DIRECTLY. INSTEAD, USE THE "AT-SCALE" BASH SCRIPT,
WHICH IS DESCRIBED HERE: https://timothy-barry.github.io/sceptre/articles/sceptre-at-scale.html
}
\details{
This function runs the gRNA precomputation on a selected "pod" of gRNAs (as identified by the pod_id).
It stores the result in the gRNA precomputation directory.
}
