% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_2.R
\name{run_power_check}
\alias{run_power_check}
\title{Run power check}
\usage{
run_power_check(
  sceptre_object,
  output_amount = 1,
  print_progress = TRUE,
  parallel = FALSE,
  n_processors = "auto",
  log_dir = tempdir()
)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}

\item{output_amount}{(optional; default \code{1}) an integer taking values 1, 2, or 3 specifying the amount of information to return}

\item{print_progress}{(optional; default \code{TRUE}) a logical indicating whether to print progress updates}

\item{parallel}{(optional; default \code{FALSE}) a logical indicating whether to run the function in parallel}

\item{n_processors}{(optional; default "auto") an integer specifying the number of processors to use if \code{parallel} is set to \code{TRUE}. The default, "auto," automatically detects the number of processors available on the machine.}

\item{log_dir}{(optional; default \code{tempdir()}) a string indicating the directory in which to write the log files (ignored if \code{parallel = FALSE})}
}
\value{
an updated \code{sceptre_object} in which the power check has been carried out
}
\description{
\code{run_power_check()} runs the power check. See \href{https://timothy-barry.github.io/sceptre-book/run-power-check-and-discovery-analysis.html}{Chapter 6 of the manual} for detailed information about this function.
}
\examples{
library(sceptredata)
data(highmoi_example_data)
data(grna_target_data_frame_highmoi)
# import data
sceptre_object <- import_data(
 response_matrix = highmoi_example_data$response_matrix,
 grna_matrix = highmoi_example_data$grna_matrix,
 grna_target_data_frame = grna_target_data_frame_highmoi,
 moi = "high",
 extra_covariates = highmoi_example_data$extra_covariates,
 response_names = highmoi_example_data$gene_names
)

# set analysis parameters, assign grnas, run qc
positive_control_pairs <- construct_positive_control_pairs(sceptre_object)
sceptre_object <- sceptre_object |>
 set_analysis_parameters(
   side = "left",
   resampling_mechanism = "permutations",
   positive_control_pairs = positive_control_pairs
 ) |>
 assign_grnas(method = "thresholding") |>
 run_qc() |>
 run_power_check()
}
