% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_2.R
\name{run_power_check}
\alias{run_power_check}
\title{Run power check}
\usage{
run_power_check(
  sceptre_object,
  output_amount = 1,
  print_progress = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}

\item{output_amount}{(optional) an integer taking values 1, 2, or 3 specifying the amount of information to return}

\item{print_progress}{(optional) a logical indicating whether to print progress updates}

\item{parallel}{(optional) a logical indicating whether to run the function in parallel}
}
\value{
an updated \code{sceptre_object} in which the power check has been carried out
}
\description{
\code{run_power_check()} runs the power check. See \href{https://timothy-barry.github.io/sceptre-book/run-power-check-and-discovery-analysis.html}{Chapter 6 of the manual} for detailed information about this function.
}
\examples{
# see example via ?sceptre
}
