% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_interface_v2.R
\name{run_sceptre}
\alias{run_sceptre}
\title{Run \code{sceptre}}
\usage{
run_sceptre(
  response_matrix,
  grna_matrix,
  covariate_data_frame,
  grna_group_data_frame,
  moi,
  formula_object,
  calibration_check,
  response_grna_group_pairs = NULL,
  control_group = "default",
  resampling_mechanism = "default",
  side = "both",
  fit_parametric_curve = TRUE,
  n_nonzero_trt_thresh = 7L,
  n_nonzero_cntrl_thresh = 7L,
  grna_assign_threshold = 5L,
  calibration_group_size = NULL,
  n_calibration_pairs = NULL,
  B1 = 499L,
  B2 = 4999L,
  B3 = 24999L,
  print_progress = TRUE,
  output_amount = 1L
)
}
\arguments{
\item{response_matrix}{(required) a matrix of raw expression counts. The responses (e.g., genes or proteins) should be in the rows, and the cells should be in the columns. The row names should be the unique IDs of the responses. The matrix can be a standard (dense) matrix or a sparse matrix of class \code{dgCMatrix}, \code{dgRMatrix}, or \code{dgTMatrix}.}

\item{grna_matrix}{(required) a matrix of gRNA expression counts. The gRNAs should be in the rows, and the cells should be in the columns. The row names should be the unique IDs of the gRNAs. The matrix can be a standard (dense) matrix or a sparse matrix of class \code{dgCMatrix}, \code{dgRMatrix}, or \code{dgTMatrix}. (See "Notes" for details about passing a gRNA matrix containing user-specified gRNA-to-cell assignments.)}

\item{covariate_data_frame}{(required) a data frame containing the cell-specific covariates (e.g., sequencing batch, total UMI count, etc.), with cells in the rows and covariates in the columns.}

\item{grna_group_data_frame}{(required) a data frame specifying the "group" to which each individual gRNA belongs. This data frame should contain columns \code{grna_id} and \code{grna_group}, with \code{grna_group} specifying the group membership of a given gRNA in the \code{grna_id} column. Typically, gRNAs that target the same site are grouped into the same gRNA group. Non-targeting gRNAs (if present) should be assigned a gRNA group label of "non-targeting".}

\item{moi}{the MOI of the dataset, either "low" or "high".}

\item{formula_object}{(required) an R formula object specifying how to adjust for the covariates in the \code{covariate_data_frame}.}

\item{calibration_check}{(required) a logical (i.e., \code{TRUE}/\code{FALSE}) value indicating whether to run a calibration check (\code{TRUE}) or a discovery analysis (\code{FALSE}). See the "Details" section for a more complete description of the calibration and discovery analyses.}

\item{response_grna_group_pairs}{(required) a data frame specifying the set of response-gRNA group pairs to test for association. The data frame should contain columns \code{response_id} and \code{grna_group}.}

\item{control_group}{(optional; moderate importance; default depends on MOI) the set of cells against which the cells that received a given targeting gRNA group are compared, either "complement" (for the complement set) or "nt" (for the cells containing a non-targeting gRNA). The only valid option for high MOI data is "complement", as few (if any) cells contain exclusively non-targeting gRNAs in high MOI. The default for low MOI data is "nt".}

\item{resampling_mechanism}{(optional; moderate importance; default depends on MOI) the resampling mechanism used to carry out inference, either "crt" (for the conditional randomization test) or "permutations" (for the permutation test). The default is to use "crt" in high MOI and "permutations" in low MOI.}

\item{side}{(optional; moderate importance; default "both") the sidedness of the test, either two-sided ("both"), left-tailed ("left"), or right-tailed ("right").}

\item{fit_parametric_curve}{(optional; moderate importance; default \code{TRUE}) a logical (i.e., \code{TRUE}/\code{FALSE}) indicating whether to fit a parametric curve to the distribution of null test statistics and to compute a p-value using this fitted parametric curve. Setting this argument to \code{TRUE} (the default) enables \code{sceptre} to return very small p-values.}

\item{n_nonzero_trt_thresh}{(optional; moderate importance; default 7) For a given response-gRNA group pair, \code{n_nonzero_trt} is the number of "treatment cells" (i.e., cells that have received the given targeting gRNA group) with nonzero expression. \code{sceptre} filters for response-gRNA group pairs with an \code{n_nonzero_trt} value equal to or greater than \code{n_nonzero_trt_thresh}.}

\item{n_nonzero_cntrl_thresh}{(optional; moderate importance; default 7) For a given response-gRNA group pair, \code{n_nonzero_cntrl} is the number of "control cells" (i.e., cells against which the treatment cells are compared) with nonzero expression. \code{sceptre} filters for pairs with a \code{n_nonzero_control} value equal to or greater than \code{n_nonzero_control_thresh}.}

\item{grna_assign_threshold}{(optional; lower importance; default 5) the threshold used to assign gRNAs to cells on high MOI data. A given gRNA is assigned to a cell if the UMI count of the gRNA within that cell is greater than or equal to the threshold. (In low MOI, this argument is ignored, as gRNAs are assigned to cells via a maximum as opposed to thresholding operation.)}

\item{calibration_group_size}{(optional; lower importance; default \code{NULL}) the number of NT gRNAs to assign to each negative control gRNA group in the calibration check. By default, \code{calibration_group_size} is set equal to the median group size of the targeting gRNA groups.}

\item{n_calibration_pairs}{(optional; lower importance; default \code{NULL}) the number of negative control pairs to analyze in the calibration check. By default, this argument is set equal to the number of discovery pairs (i.e., pairs specified in the data frame \code{response_grna_group_pairs}) that passes pairwise QC.}

\item{B1}{(optional; lower importance; default 499) the number of null test statistics to compute in the first stage of the permutation test.}

\item{B2}{(optional; lower importance; default 4999) the number of null test statistics to compute the second stage of the permutation test.}

\item{B3}{(optional; lower importance; default 24999) the number of null test statistics to compute the third stage of the permutation test.}

\item{print_progress}{(optional; lower importance; default TRUE) a logical (i.e., \code{TRUE}/\code{FALSE} value) indicating whether to print the progress of the function.}

\item{output_amount}{(optional; lower importance; default 1) an integer specifying the amount of information to return as part of the output, either 1 (least), 2 (intermediate), or 3 (most). (See "Value".)}
}
\value{
A data frame containing the following columns: \code{response_id}, \code{grna_group}, \code{n_nonzero_trt}, \code{n_nonzero_cntrl}, and \code{p_value}.
\itemize{
\item{\code{response_id}}: the ID of the response
\item{\code{grna_group}}: the gRNA group against which the response is tested. When running a calibration check, \code{grna_group} is a negative control gRNA group constructed by grouping together individual NT gRNAs. When running a discovery analysis, by contrast, \code{grna_group} is a targeting gRNA group.
\item{\code{n_nonzero_trt}}: the number of treatment cells (i.e., cells that have received a targeting gRNA) with nonzero expression
\item{\code{n_nonzero_cntrl}}: the number of control cells (i.e., cells that have received a non-targeting gRNA) with nonzero expression
\item{\code{p_value}}: the \code{sceptre} p-value
\item{\code{log_2_fold_change}}: the \code{sceptre}-estimated log (base 2) fold change in expression, controlling for the covariates
}
Rows are ordered according to \code{p_value}. Response-gRNA group pairs that fail to pass pairwise QC are assigned a p-value and log fold change estimate of \code{NA}. When \code{output_amount} is set to 2, the following additional columns are returned:

\itemize{
\item{\code{stage}}: the stage at which the p-value was computed, one of 1, 2, or 3 (See Notes)
\item{\code{z_orig}}: the original z-score computed on the raw data
\item{\code{xi}, \code{omega}, \code{alpha}}: the fitted parameters of the skew-normal distribution. (These columns are set to \code{NA} if a parametric curve was not fitted to the null distribution of test statistics.)
}

When \code{output_amount} is set to 3, the resampling distribution of the test statistics also is returned, stored in the columns \code{z_null_1}, \code{z_null_2}, \code{z_null_3}, ... .
}
\description{
Applies \code{sceptre} to analyze a single-cell CRISPR screen dataset.
}
\details{
A typical \code{sceptre} analysis consists of three main steps. First, the user sets up the analysis, which entails specifying (i) the response expression matrix, (ii) the gRNA expression matrix, (iii) the cell covariate data frame, (iv) the gRNA group data frame, (v) the formula object, and (vi) the set of response-gRNA group pairs to test for association. Second, the user carries out the calibration check by passing the above objects to the function \code{run_sceptre}, setting \code{calibration_check} to \code{TRUE}. In carrying out the calibration check, \code{sceptre} automatically constructs a set of negative control response-gRNA group pairs to test for association; these pairs are "matched" to the discovery pairs in several respects. The user verifies calibration of the negative control p-values by plotting the calibration results via \code{plot_calibration_result}. Finally, the user runs the discovery analysis by again calling the function \code{run_sceptre}, this time setting \code{calibration_check} to \code{FALSE}. The user visualizes the discovery results via calls to \code{compare_calibration_and_discovery_results} and \code{make_volcano_plot} and obtains the final set of discoveries via a call to \code{obtain_discovery_set}.
}
\note{
The notes are arranged roughly in order of most important to least important.

\itemize{
\item{The function \code{compute_cell_covariates} can be used to help compute the \code{covariate_data_frame}, and the function \code{generate_all_pairs} can be used to help compute \code{response_grna_group_pairs}.}
\item{When constructing the \code{formula_object}, it is best practice to log-transform the integer count-based variables (e.g., total UMI count, number of expressed responses).}
\item{By default, \code{sceptre} assigns gRNAs to cells for users. Users instead may assign gRNAs to cells themselves, passing a logical (i.e., \code{TRUE}/\code{FALSE})  \code{grna_matrix} to \code{sceptre}. The logical matrix should be a standard (dense) R matrix or a sparse matrix of type \code{lgCMatrix}, \code{lgRMatrix}, \code{lgTMatrix}.} In low MOI each column should contain a single \code{TRUE} value, which indicates which gRNA has infected the given cell.
\item{\code{sceptre} tests a given pair for association in three stages. First, \code{sceptre} computes \code{B1} null statistics and calculates an initial empirical p-value \code{p1} using these null statistics. If \code{p1} is promising (i.e., \code{p1} < 0.02), then \code{sceptre} proceeds to stage 2; otherwise, \code{sceptre} returns \code{p1}. In stage 2, \code{sceptre} computes \code{B2} null statistics and fits a parametric curve (by default, a skew normal distribution) to these null statistics. If the fit of the parametric curve is good, then \code{sceptre} returns \code{p2}. Otherwise, \code{sceptre} proceeds to stage 3. In stage 3, \code{sceptre} computes \code{B3} null statistics and calculates an empirical p-value \code{p3} using these null statistics. \code{sceptre} then returns \code{p3}. When \code{resampling_mechanism} is set to "crt" (the default for high MOI data), the stage 3 empirical p-value is computed using the null test statistics from stage 2 (to save compute).}
}
}
