% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_method_functions.R
\name{run_sceptre_gRNA_gene_pair}
\alias{run_sceptre_gRNA_gene_pair}
\title{Run SCEPTRE on a gRNA-gene pair}
\usage{
run_sceptre_gRNA_gene_pair(
  gene_expressions,
  gRNA_expressions,
  covariate_matrix,
  side = "both",
  B = 1500,
  full_output = FALSE,
  seed = 4
)
}
\description{
Runs SCEPTRE on a given gRNA-gene pair.
}
\examples{
# load the example data
data(gene_matrix); data(gRNA_matrix); data(covariate_matrix)
gene_expressions <- gene_matrix[1,]
gRNA_expressions <- gRNA_matrix[1,]
# run method
run_sceptre_gRNA_gene_pair(gene_expressions, gRNA_expressions, covariate_matrix, "left")
}
