% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_method_functions.R
\name{run_sceptre_gRNA_gene_pair}
\alias{run_sceptre_gRNA_gene_pair}
\title{Run SCEPTRE on a gRNA-gene pair}
\usage{
run_sceptre_gRNA_gene_pair(
  expressions,
  gRNA_indicators,
  covariate_matrix,
  side = "left",
  gene_precomp_size = NULL,
  B = 500,
  seed = NULL,
  reduced_output = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{expressions}{a vector a gene expressions}

\item{gRNA_indicators}{a vector of gRNA inicators}

\item{covariate_matrix}{the data frame of cell-specific covariates}

\item{side}{(optional; default "left") side of the test; one of "left," "right," and "both."}

\item{gene_precomp_size}{(optional) the pre-computed size of the gene NB distribution; if not supplied, will be estimated}

\item{B}{(optional; default 500) number of conditional randomization test resamples}

\item{seed}{(optional; default 4) seed to the random number generator}

\item{reduced_output}{(optional; default TRUE) return the reduced output?}

\item{verbose}{(optional; default TRUE) print status updates?}
}
\value{
If reduced_output is TRUE, a data frame containing the following columns: p-value, skew-t fit success, skew-t fit MLEs (xi, omega, alpha, nu), the original signed test statistic, and the number of successful resamples; if reduced_output is FALSE, return a list containing all of the above, plus the vector of resampled test statistics.
}
\description{
Runs SCEPTRE on a given gRNA-gene pair. The function requires as arguments the gene expression vector, the gRNA indicator vector, and the covariate matrix.
}
\examples{
\dontrun{
data(expressions)
data(gRNA_indicators)
data(covariate_matrix)
result <- run_sceptre_gRNA_gene_pair(expressions,
gRNA_indicators,
covariate_matrix)
}
}
