% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_method_functions.R
\name{run_sceptre_gRNA_gene_pair}
\alias{run_sceptre_gRNA_gene_pair}
\title{Run sceptre on a gRNA-gene pair}
\usage{
run_sceptre_gRNA_gene_pair(
  expressions,
  gRNA_indicators,
  covariate_matrix,
  gene_precomp_size = NULL,
  B = 500,
  seed = NULL,
  side = "left",
  reduced_output = TRUE
)
}
\arguments{
\item{expressions}{a vector a gene expressions}

\item{gRNA_indicators}{a vector of gRNA inicators}

\item{covariate_matrix}{the matrix of cell-specific covariates (e.g., library size, batch effect, cell cycle, etc.)}

\item{gene_precomp_size}{(optional) the pre-computed size of the gene NB distribution}

\item{B}{number of resamples (default 500)}

\item{seed}{(optional) seed to the random number generator}

\item{side}{an argument to set test for left-sided, right-sided or both-sided. Default as 'left' and can take 'left', 'right' or 'both'}
}
\value{
a p-value of the null hypothesis of no gRNA effect on gene expression
}
\description{
This function runs the sceptre algorithm on a single gRNA-gene pair. It requires as arguments the gene expression vector, the gRNA indicator vector, and the covariate matrix. Users optionally can pass the gRNA precomputation or gene precomputation as arguments.
}
\examples{
offsite_dir <- "/Volumes/tims_new_drive/research/sceptre_files"
source("/Users/timbarry/Box/SCEPTRE/sceptre_paper/analysis_drivers_xie/sceptre_function_args.R")
expressions <- cell_gene_expression_matrix[,2][cell_subset]
gRNA_indicators <- (read.fst(gRNA_indicator_matrix_fp, columns = "chr5:54325645-54326045") \%>\% pull() \%>\% as.integer())[cell_subset]
covariate_matrix <- if (nrow(covariate_matrix) == 106666) covariate_matrix else covariate_matrix[cell_subset,]
run_sceptre_gRNA_gene_pair(expressions, gRNA_indicators, covariate_matrix)

}
