% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_interface_v2.R
\name{run_sceptre_highmoi}
\alias{run_sceptre_highmoi}
\title{Run \code{sceptre} (high MOI)}
\usage{
run_sceptre_highmoi(
  response_matrix,
  grna_matrix,
  covariate_data_frame,
  grna_group_data_frame,
  formula_object,
  calibration_check,
  response_grna_group_pairs = NULL,
  control_group = "default",
  resampling_mechanism = "default",
  n_nonzero_trt_thresh = 7L,
  n_nonzero_cntrl_thresh = 7L,
  side = "both",
  grna_assign_threshold = 5L,
  output_amount = 1L,
  fit_skew_normal = TRUE,
  calibration_group_size = NULL,
  n_calibration_pairs = NULL,
  B1 = 499L,
  B2 = 4999L,
  B3 = 24999L,
  print_progress = TRUE
)
}
\description{
Run \code{sceptre} (high MOI)
}
\examples{
\dontrun{
library(Matrix)

data(response_matrix_highmoi) # response-by-cell expression matrix
data(grna_matrix_highmoi) # gRNA-by-cell expression matrix
data(covariate_data_frame_highmoi) # cell-by-covariate data frame
data(grna_group_data_frame_highmoi) # gRNA group information
data(discovery_pairs_highmoi)

# 1. set the pairs to analyze
response_grna_group_pairs <- discovery_pairs_highmoi

# 2. set the formula object
formula_object <- formula(~log(grna_n_umis) + log(gene_n_umis) + batch + p_mito)

# 3. automatic calibration check
calibration_result <- run_sceptre_highmoi(response_matrix = response_matrix_highmoi,
grna_matrix = grna_matrix_highmoi,
covariate_data_frame = covariate_data_frame_highmoi,
grna_group_data_frame = grna_group_data_frame_highmoi,
response_grna_group_pairs = discovery_pairs_highmoi,
formula_object = formula_object,
calibration_check = TRUE)

# 4. run discovery analysis
discovery_result <- run_sceptre_highmoi(response_matrix = response_matrix_highmoi,
grna_matrix = grna_matrix_highmoi,
covariate_data_frame = covariate_data_frame_highmoi,
grna_group_data_frame = grna_group_data_frame_highmoi,
response_grna_group_pairs = discovery_pairs_highmoi,
formula_object = formula_object,
calibration_check = FALSE)
}
}
