% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_memory_sceptre.R
\name{run_sceptre_in_memory}
\alias{run_sceptre_in_memory}
\title{Run \code{sceptre} in memory}
\usage{
run_sceptre_in_memory(
  gene_matrix,
  gRNA_matrix,
  covariate_matrix,
  gene_gRNA_pairs,
  side = "both",
  storage_dir = tempdir(),
  regularization_amount = 0.1,
  B = 1000,
  full_output = FALSE,
  seed = 4
)
}
\arguments{
\item{gene_matrix}{a gene-by cell expression matrix; the rows (i.e., gene IDs) should be named}

\item{gRNA_matrix}{a gRNA-by cell expression matrix; the rows (i.e., gRNA IDs) should be named}

\item{covariate_matrix}{the cell-specific matrix of technical factors, ideally containing the following covariates: log-transformed gene library size, log-transformed gRNA library size, percent mitochondrial reads, and batch}

\item{gene_gRNA_pairs}{a data frame specifying the gene-gRNA pairs to test for association; the data frame should contain columns named \code{gene_id} and \code{gRNA_id}}

\item{side}{sidedness of the test; one of "both," "left," and "right"}

\item{storage_dir}{directory in which to store the intermediate computations}

\item{regularization_amount}{non-negative number specifying the amount of regularization to apply to the negative binomial dispersion parameter estimates}

\item{B}{number of resamples to draw for the conditional randomization test}

\item{full_output}{return the full output (TRUE) or a simplified, reduced output (FALSE)?}

\item{seed}{seed to the random number generator}
}
\value{
a data frame containing columns \code{gene_id}, \code{gRNA_id}, \code{p_value}, and \code{z_value}. See
}
\description{
The core function of the \code{sceptre} package. This function tests for association between a set of gRNAs and a set of genes, returning a p-value for each pairwise test of association.
}
\examples{
# 1. load the data
data(gene_matrix) # i. gene expression matrix
data(gRNA_matrix) # ii. gRNA expression matrix
data(covariate_matrix) # iii. covariate matrix
data(gRNA_grps) # iv. gRNAs grouped by target site
data(gene_gRNA_pairs) # v. gene-gRNA pairs to analyze
# 2. (Optional) group together gRNAs that target the same site
gRNA_matrix_grouped <- combine_gRNAs(gRNA_matrix, gRNA_grps)
# 3. run method (takes ~40s on an 8-core Macbook Pro)
result <- run_sceptre_in_memory(gene_matrix, gRNA_matrix_grouped, covariate_matrix, gene_gRNA_pairs)
}
