% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceptre.R
\docType{package}
\name{sceptre}
\alias{sceptre}
\title{sceptre}
\description{
\code{sceptre} is a resource-light, fast, and statistically principled software for single-cell CRISPR screen analysis. \code{sceptre} is powered by several methodological and algorithmic advances in robust differential expression analysis and fast permutation testing. \code{sceptre} has modules for low multiplicity of infection (MOI) and high MOI single-cell CRISPR screen analysis.
}
