% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceptre.R
\docType{package}
\name{sceptre}
\alias{sceptre}
\title{sceptre}
\description{
SCEPTRE (analysis of single cell perturbation screens via conditional resampling) is a method for single-cell CRISPR screen analysis. SCEPTRE infers gene-perturbation associations by modeling the random assortment of CRISPR guide RNAs among cells instead of modeling gene expression, thereby remaining valid despite confounder presence and arbitrary misspecification of the gene expression model.
}
