% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_1.R
\name{set_analysis_parameters}
\alias{set_analysis_parameters}
\title{Set analysis parameters}
\usage{
set_analysis_parameters(
  sceptre_object,
  discovery_pairs,
  positive_control_pairs = data.frame(grna_target = character(0), response_id =
    character(0)),
  side = "both",
  grna_grouping_strategy = "union",
  formula_object = "default",
  fit_parametric_curve = TRUE,
  control_group = "default",
  resampling_mechanism = "default",
  multiple_testing_method = "BH",
  multiple_testing_alpha = 0.1
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{discovery_pairs}{TBD}

\item{positive_control_pairs}{TBD}

\item{side}{TBD}

\item{grna_grouping_strategy}{TBD}

\item{formula_object}{TBD}

\item{fit_parametric_curve}{TBD}

\item{control_group}{TBD}

\item{resampling_mechanism}{TBD}

\item{multiple_testing_method}{TBD}

\item{multiple_testing_alpha}{TBD}

\item{B1}{TBD}

\item{B2}{TBD}

\item{B3}{TBD}
}
\description{
Set analysis parameters
}
