% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_1.R
\name{set_analysis_parameters}
\alias{set_analysis_parameters}
\title{Set analysis parameters}
\usage{
set_analysis_parameters(
  sceptre_object,
  discovery_pairs,
  positive_control_pairs = data.frame(),
  formula_object = "default",
  side = "both",
  fit_parametric_curve = TRUE,
  control_group = "default",
  resampling_mechanism = "default",
  B1 = 499L,
  B2 = 4999L,
  B3 = 24999L,
  multiple_testing_method = "BH",
  multiple_testing_alpha = 0.1
)
}
\arguments{
\item{sceptre_object}{TBD}

\item{discovery_pairs}{TBD}

\item{positive_control_pairs}{TBD}

\item{formula_object}{TBD}

\item{side}{TBD}

\item{fit_parametric_curve}{TBD}

\item{control_group}{TBD}

\item{resampling_mechanism}{TBD}

\item{B1}{TBD}

\item{B2}{TBD}

\item{B3}{TBD}

\item{multiple_testing_method}{TBD}

\item{multiple_testing_alpha}{TBD}
}
\description{
Set analysis parameters
}
