% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_1.R
\name{set_analysis_parameters}
\alias{set_analysis_parameters}
\title{Set analysis parameters}
\usage{
set_analysis_parameters(
  sceptre_object,
  discovery_pairs = data.frame(grna_target = character(0), response_id = character(0)),
  positive_control_pairs = data.frame(grna_target = character(0), response_id =
    character(0)),
  side = "both",
  grna_integration_strategy = "union",
  formula_object = "default",
  resampling_approximation = "skew_normal",
  control_group = "default",
  resampling_mechanism = "default",
  multiple_testing_method = "BH",
  multiple_testing_alpha = 0.1
)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}

\item{discovery_pairs}{(optional) a data frame with columns \code{grna_target} and \code{response_id} specifying the discovery pairs to analyze}

\item{positive_control_pairs}{(optional) a data frame with columns \code{grna_target} and \code{response_id} specifying the positive control pairs to analyze}

\item{side}{(optional; default \code{"both"}) the sidedness of the test, one of \code{"left"}, \code{"right"}, or \code{"both"}}

\item{grna_integration_strategy}{(optional; default \code{"union"}) a string specifying the gRNA integration strategy, either \code{"singleton"}, \code{"union"}, or \code{"bonferroni"}}

\item{formula_object}{(optional) a formula object specifying how to adjust for the covariates in the model}

\item{resampling_approximation}{(optional; default \code{"skew_normal"}) a string indicating the resampling approximation to make to the null distribution of test statistics, either \code{"skew_normal"} or \code{"no_approximation"}}

\item{control_group}{(optional) a string specifying the control group to use in the differential expression analysis, either \code{"complement"} or \code{"nt_cells"}}

\item{resampling_mechanism}{(optional) a string specifying the resampling mechanism to use, either \code{"permutations"} or \code{"crt"}}

\item{multiple_testing_method}{(optional; default \code{"BH"}) a string specifying the multiple testing correction method to use; see \code{p.adjust.methods} for options}

\item{multiple_testing_alpha}{(optional; default \code{0.1}) a numeric specifying the nominal level of the multiple testing correction method}
}
\value{
an updated \code{sceptre_object} in which the analysis parameters have been set
}
\description{
\code{set_analysis_parameters()} sets the analysis parameters that control how the statistical analysis is to be conducted. See \href{https://timothy-barry.github.io/sceptre-book/set-analysis-parameters.html}{Chapter 2 of the manual} for more detailed information about this function.
}
\note{
Every argument to this function is optional, but typically, users want to specify \code{discovery_pairs} at minimum.
}
\examples{
data(highmoi_example_data)
data(grna_target_data_frame_highmoi)
# import data
sceptre_object <- import_data(
  response_matrix = highmoi_example_data$response_matrix,
  grna_matrix = highmoi_example_data$grna_matrix,
  grna_target_data_frame = grna_target_data_frame_highmoi,
  moi = "high",
  extra_covariates = highmoi_example_data$extra_covariates,
  response_names = highmoi_example_data$gene_names
)

# set analysis parameters
positive_control_pairs <- construct_positive_control_pairs(sceptre_object)
discovery_pairs <- construct_cis_pairs(sceptre_object,
  positive_control_pairs = positive_control_pairs,
  distance_threshold = 5e6
)
sceptre_object <- sceptre_object |>
  set_analysis_parameters(
    discovery_pairs = discovery_pairs,
    positive_control_pairs = positive_control_pairs,
    side = "left"
  )
}
