% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_1.R
\name{set_analysis_parameters}
\alias{set_analysis_parameters}
\title{Set analysis parameters}
\usage{
set_analysis_parameters(
  sceptre_object,
  discovery_pairs,
  positive_control_pairs = data.frame(grna_target = character(0), response_id =
    character(0)),
  side = "both",
  grna_integration_strategy = "union",
  formula_object = "default",
  fit_parametric_curve = TRUE,
  control_group = "default",
  resampling_mechanism = "default",
  multiple_testing_method = "BH",
  multiple_testing_alpha = 0.1
)
}
\arguments{
\item{sceptre_object}{a \code{sceptre_object}}

\item{discovery_pairs}{a data frame with columns \code{grna_target} and \code{response_id} specifying the discovery pairs to analyze}

\item{positive_control_pairs}{(optional) a data frame with columns \code{grna_target} and \code{response_id} specifying the positive control pairs to analyze}

\item{side}{(optional; default \code{"both"}) the sidedness of the test, one of \code{"left"}, \code{"right"}, or \code{"both"}}

\item{grna_integration_strategy}{(optional; default \code{"union"}) a string specifying the gRNA integration strategy, either \code{"singleton"} or \code{"union"}}

\item{formula_object}{(optional) a formula object specifying how to adjust for the covariates in the model}

\item{fit_parametric_curve}{(optional; default \code{TRUE}) a logical indicating whether to fit a parametric curve to the null distribution of test statistics}

\item{control_group}{(optional) a string specifying the control group to use, either \code{"complement"} or \code{"nt_cells"}}

\item{resampling_mechanism}{(optional) a string specifying the resampling mechanism to use, either \code{"permutations"} or \code{"crt"}}

\item{multiple_testing_method}{(optional; default \code{"BH"}) a string specifying the multiple testing correction method to use; see \code{p.adjust.methods} for options}

\item{multiple_testing_alpha}{(optional; default \code{0.1}) a numeric specifying the nominal level of the multiple testing correction method}
}
\value{
an updated \code{sceptre_object} in which the analysis parameters have been set
}
\description{
\code{set_analysis_parameters()} sets the analysis parameters that control how the statistical analysis is to be conducted. See \href{https://timothy-barry.github.io/sceptre-book/set-analysis-parameters.html}{Chapter 2 of the manual} for detailed information about this function.
}
\examples{
# see example via ?sceptre
}
