% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_crispr_screen_data}
\alias{simulate_crispr_screen_data}
\title{Simulate CRISPR screen data}
\usage{
simulate_crispr_screen_data(
  n_cells,
  grna_mean_prob,
  mRNA_mean_expression,
  covariate_effects_gRNA,
  covariate_effects_gene,
  zero_inflation,
  neg_binom_size,
  seed = NULL
)
}
\arguments{
\item{n_cells}{number of cells}

\item{grna_mean_prob}{mean probability of gRNA presence across enhancers (absent covariates)}

\item{mRNA_mean_expression}{mean expression level of gene across genes (absent covariates)}

\item{covariate_effects_gRNA}{the effect of the technical covariate on gRNA presence probability; length equals number of enhancers}

\item{covariate_effects_gene}{the effect of the technical covariate on mean gene expression; length equals number of genes}

\item{zero_inflation}{zero inflation rate (between 0 and 1)}

\item{neg_binom_size}{size of the negative binomial distributions for gene expression}

\item{seed}{(optonal) seed to the random number generator}
}
\value{
list with components (i) cell_by_gene_expressions, (ii) cell_by_enhancer_perturbation_indicators, and (iii) Z
}
\description{
Simulate CRISPR screen data
}
\examples{
r <- simulate_crispr_screen_data(n_cells = 1000,
grna_mean_prob = 0.2,
mRNA_mean_expression = 40,
covariate_effects_gRNA = rep(4, 10),
covariate_effects_gene = rep(2, 50),
zero_inflation = 0,
neg_binom_size = 2
)
}
