% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_analysis_functs_2.R
\name{write_outputs_to_directory}
\alias{write_outputs_to_directory}
\title{Write outputs to directory}
\usage{
write_outputs_to_directory(sceptre_object, directory)
}
\arguments{
\item{sceptre_object}{TBD}

\item{directory}{TBD}
}
\description{
Write outputs to directory
}
